\name{adapr-package}
\alias{adapr-package}
\alias{adapr}
\docType{package}
\title{
\packageTitle{adapr}
}
\description{
\packageDescription{adapr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{adapr}
\packageIndices{adapr}
Tracks reading and writing within R scripts that are organized into a directed acyclic graph
}
\author{
\packageAuthor{adapr}

Maintainer: \packageMaintainer{adapr}
}
\references{
None.
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
#Digest()
}
