% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Library.R
\name{adaprDependencies}
\alias{adaprDependencies}
\title{Returns character string of adapr R package import dependencies}
\usage{
adaprDependencies()
}
\value{
character vector of package names
}
\description{
Returns character string of adapr R package import dependencies
}
\examples{
\dontrun{
adaprDependencies()
} 

}
