% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_expand.2.R
\name{adaprHomeDir}
\alias{adaprHomeDir}
\title{Identify adapr Home directory. Location of options file and project listing.}
\usage{
adaprHomeDir()
}
\value{
Full filepath to Adapr options and project listings directory.
}
\description{
Identify adapr Home directory. Location of options file and project listing.
}
\details{
This is automatically handled by defaultAdaprSetup(), but can be controlled in an "R profile" by adding adaprHomeDir R option ( e.g.,  options(adaprHomeDir="myPath")).
}
\examples{
\dontrun{
 adaprHomeDir()
} 

}
