% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaprInstall.R
\name{adaprInstall}
\alias{adaprInstall}
\title{Install adapr in a library
adaprInstall.R}
\usage{
adaprInstall(library.location = getProjectLibrary("adaprHome"),
  betaTF = getAdaprOptions()$adaprBeta)
}
\arguments{
\item{library.location}{file path to library}

\item{betaTF}{logical indicating whether to install from github or CRAN.}
}
\value{
logical for succesful creation or not
}
\description{
Install adapr in a library
adaprInstall.R
}
\details{
Installs dependencies
}
\examples{
\dontrun{
 adaprInstall(library.location=getProjectLibrary("adaprHome") ,betaTF=FALSE)
} 

}
