% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapr_update.R
\name{adaprUpdate}
\alias{adaprUpdate}
\title{Updates latest adapr from gitHub.}
\usage{
adaprUpdate()
}
\value{
message from git_
}
\description{
Updates latest adapr from gitHub.
}
\examples{
\dontrun{
adaprUpdate()
} 
}
