% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_library.R
\name{add_package}
\alias{add_package}
\title{Add R package to a project}
\usage{
add_package(project.id = get.project(), library.name, library.install = NA,
  library.specific = FALSE)
}
\arguments{
\item{project.id}{project.id to add R package to}

\item{library.name}{R package name to add}

\item{library.install}{Command to install library: NA for CRAN, bioC for bioconductor}

\item{library.specific}{logical indicate whether package is for specific R script}
}
\value{
Library information data
}
\description{
Add R package to a project
}
\details{
Installs and loads all packages
}
\examples{
\dontrun{
add_package("adaprHome","ggplot2")
} 

}
