% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_library.R
\name{addPackage}
\alias{addPackage}
\title{Add R package to a project}
\usage{
addPackage(project.id = getProject(), library.name, library.install = NA,
  library.specific = FALSE)
}
\arguments{
\item{project.id}{project.id to add R package to}

\item{library.name}{R package name to add}

\item{library.install}{Command to install library: NA for CRAN, bioC for bioconductor}

\item{library.specific}{logical indicate whether package is for specific R script}
}
\value{
Library information data
}
\description{
Add R package to a project
}
\details{
Not for direct use. Installs and loads all packages
}
\examples{
\dontrun{
addPackage("adaprHome","ggplot2")
} 

}
