% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_branch.R
\name{AppLoadFlex}
\alias{AppLoadFlex}
\title{Loads a single R object from file for a R Shiny app}
\usage{
AppLoadFlex(project.id = getProject(), path, file, read.fcn = readRDS, ...)
}
\arguments{
\item{project.id}{project name from which to load file}

\item{path}{directory that contains file to be loaded}

\item{file}{contains R object}

\item{read.fcn}{function to read the file, default readRDS}

\item{...}{arguments passed to read.fcn}
}
\value{
object for file that was read
}
\description{
Loads a single R object from file for a R Shiny app
}
\examples{
\dontrun{
processed <- AppLoadFlex("adaprTest","Results/read_data.R","cardata.RData")
} 
}
