% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aWrite.R
\name{arcRead}
\alias{arcRead}
\title{Loads R object from archivist repository within the results directory of another adapr Script}
\usage{
arcRead(rscript = "read_data.R", description = "xyplot",
  project.id = getProject())
}
\arguments{
\item{rscript}{name of R script that loaded the function}

\item{description}{character description. Need for access with arcRead}

\item{project.id}{project to search within}
}
\description{
Loads R object from archivist repository within the results directory of another adapr Script
}
\details{
For use within R adapr script. This complements the use of arcWrite. The description should match the description in arcWrite.
}
\examples{
\dontrun{
# In Script that writes the archivist object:
# arcWrite(myplot,"xyplot")

# In Script that reads the object:
processed <- arcRead("read_data.R","xyplot")
} 


}
