% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aWrite.R
\name{arcWrite}
\alias{arcWrite}
\title{Saves R object to archivist repository in results directory of adapr Script}
\usage{
arcWrite(Robj, description)
}
\arguments{
\item{Robj}{R object to be written}

\item{description}{character description. Need for access with aRead}
}
\value{
Value of object description
}
\description{
Saves R object to archivist repository in results directory of adapr Script
}
\details{
For use within R adapr script. Makes tags in archivist 'source:rscript' and 'aName:description'. Should be paired with arcRead().
}
\examples{
\dontrun{
processed <- arcWrite(rnorm(100),"100 Gaussians")
} 


}
