% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Check_file_hash.R
\name{Check.file.hash}
\alias{Check.file.hash}
\title{Checks the consistency of the dependency directory with the files within the file system}
\usage{
Check.file.hash(dependency.dir = NULL, dependency.object = NULL)
}
\arguments{
\item{dependency.dir}{Directory with dependency information files}

\item{dependency.object}{data frame with dependency information}
}
\value{
list of information about file hase mismatches
}
\description{
Checks the consistency of the dependency directory with the files within the file system
}
\details{
Only needs one or the other argument
}

