% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_file_hash_source.R
\name{Check.file.hash.source}
\alias{Check.file.hash.source}
\title{Checks the consistency of the dependency directory with the files within the file system
Reports the source scripts that need to be updated!}
\usage{
Check.file.hash.source(dependency.dir = NULL, dependency.object = NULL)
}
\arguments{
\item{dependency.dir}{Directory with dependency information files}

\item{dependency.object}{data frame with dependency information}
}
\value{
list of information about file hase mismatches
}
\description{
Checks the consistency of the dependency directory with the files within the file system
Reports the source scripts that need to be updated!
}
\details{
Only needs one or the other argument.
}
\examples{
\dontrun{

Check.file.hash.source(pull_source_info("adaprHome")$dependency.dir)
} 

}
