% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_file_hash.R
\name{checkFileHash}
\alias{checkFileHash}
\title{Checks the consistency of the dependency directory with the files within the file system}
\usage{
checkFileHash(dependency.dir = NULL, dependency.object = NULL)
}
\arguments{
\item{dependency.dir}{Directory with dependency information files}

\item{dependency.object}{Data frame with dependency information}
}
\value{
list of information about file hase mismatches
}
\description{
Checks the consistency of the dependency directory with the files within the file system
}
\details{
Not for direct use. Only needs one or the other argument.
}
\examples{
\dontrun{
checkFileHash(pullSourceInfo("adaprHome")$dependency.dir)
} 


}
