% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_file_mtime_source.R
\name{checkFileMtimeSource}
\alias{checkFileMtimeSource}
\title{Checks the consistency of the dependency directory with the files within the file system
Reports the source scripts that need to be updated!}
\usage{
checkFileMtimeSource(dependency.dir = NULL, dependency.object = NULL)
}
\arguments{
\item{dependency.dir}{Directory with dependency information files}

\item{dependency.object}{data frame with dependency information}
}
\value{
list of information about file hase mismatches
}
\description{
Checks the consistency of the dependency directory with the files within the file system
Reports the source scripts that need to be updated!
}
\details{
Only needs one or the other argument.
}
\examples{
\dontrun{
 checkFileMtimeSource(pullSourceInfo("adaprHome")$dependency.dir)
} 

}
