% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_loader.R
\name{checkRmdMode}
\alias{checkRmdMode}
\title{Checks whether interactive R markdown session is ON}
\usage{
checkRmdMode(changeOption = FALSE)
}
\arguments{
\item{changeOption}{logical to print out and change the Rmdstart option}
}
\value{
value TRUE if R session in in interactive R markdown mode
}
\description{
Checks whether interactive R markdown session is ON
}
\examples{
\dontrun{
 checkRmdMode()
} 

}
