% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_install_library_file.R
\name{checkVersion}
\alias{checkVersion}
\title{Check install of package of specific version}
\usage{
checkVersion(package0, version0 = "", versionCheck = FALSE,
  lib = .libPaths()[1])
}
\arguments{
\item{package0}{name of package}

\item{version0}{package version}

\item{versionCheck}{logical to install specific version}

\item{lib}{path to local library}
}
\value{
logical on installed status
}
\description{
Check install of package of specific version
}
