% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witchDoc.R
\name{codeChunkDocId}
\alias{codeChunkDocId}
\title{Generate code to render creation compute time docID}
\usage{
codeChunkDocId()
}
\value{
Character vector for code chunk for computing docID
}
\description{
Generate code to render creation compute time docID
}
\examples{

# Requires internet connection to access random number generator.
\dontrun{
codeChunkDocId()
}

}
