% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_project.R
\name{commit.project}
\alias{commit.project}
\title{Git commit of project.}
\usage{
commit.project(project.id = get.project(), commit.message = "")
}
\arguments{
\item{project.id}{project to graph}

\item{commit.message}{message describing edits}
}
\value{
commit message
}
\description{
Git commit of project.
}
\details{
Need git option active. Uses git2r package.
}
\examples{
\dontrun{
commit.project("adaprHome","Did I change something?")
} 

}

