% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_project.R
\name{commit2char}
\alias{commit2char}
\title{git2r commit class to character converter}
\usage{
commit2char(commitclass)
}
\arguments{
\item{commitclass}{Commit object}
}
\value{
commit message
}
\description{
git2r commit class to character converter
}
\details{
Uses git2r package.
}
\examples{
\dontrun{
committed <- git2r::commit(repo,message =commit.message)
out <- paste("Git",commit2char(committed))
} 

}
