% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_project.R
\name{commitProject}
\alias{commitProject}
\title{Git commit of project.}
\usage{
commitProject(commit.message = "", project.id = getProject())
}
\arguments{
\item{commit.message}{message describing edits}

\item{project.id}{project to commit}
}
\value{
commit message
}
\description{
Git commit of project.
}
\details{
Need git option active. Uses git2r package.
}
\examples{
\dontrun{
commitProject("adaprHome","Did I change something?")
} 

}
