% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Condense_file_info.R
\name{Condense.file.info}
\alias{Condense.file.info}
\title{Generate condensed information about files from dependency object}
\usage{
Condense.file.info(trees)
}
\arguments{
\item{trees}{dependency object}
}
\value{
data frame with information about each file in project
}
\description{
Generate condensed information about files from dependency object
}
\examples{
\dontrun{
 trees <- Harvest.trees(pull_source_info("adaprHome")$dependency.dir)
Condense.file.info(trees)
} 

}
