% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Condense_file_info.R
\name{condenseFileInfo}
\alias{condenseFileInfo}
\title{Generate condensed information about files from dependency object}
\usage{
condenseFileInfo(trees)
}
\arguments{
\item{trees}{dependency object}
}
\value{
data frame with information about each file in project
}
\description{
Generate condensed information about files from dependency object
}
\examples{
\dontrun{
 trees <- readDependency(pullSourceInfo("adaprHome")$dependency.dir)
condenseFileInfo(trees)
} 

}
