% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyProject.R
\name{copyProject}
\alias{copyProject}
\title{Copies project to a different directory}
\usage{
copyProject(project.id = getProject(),
  targetDirectory = getProjectPublishPath(project.id), dataCopy = TRUE,
  programsOnly = FALSE, speedCopy = TRUE)
}
\arguments{
\item{project.id}{character string name of project}

\item{targetDirectory}{character string specifying which directory to copy to}

\item{dataCopy}{logicial specifying whether to copy data}

\item{programsOnly}{logical indicating to copy only program folder}

\item{speedCopy}{logical indicating to not copy if file name, mod time, and file size are identical in target directory.}
}
\value{
data frame
}
\description{
Copies project to a different directory
}
\details{
Defaults to copy to the publish directory
}
