% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_markdown.R
\name{create_markdown}
\alias{create_markdown}
\title{Renders and Rmarkdown file}
\usage{
create_markdown(target.file = paste0(source_info$file$file, "md"),
  target.dir = source_info$analysis.dir, style = "html_document",
  description = "Markdown", si, overwrite = FALSE)
}
\arguments{
\item{target.file}{Markdown file to create}

\item{target.dir}{Directory to send target file}

\item{description}{Markdown description}

\item{si}{source_info object for tracking}

\item{overwrite}{overwrite existing R markdown?}

\item{tyle}{Markdown target style}
}
\value{
File information
}
\description{
Renders and Rmarkdown file
}
\details{
Uses rmarkdown library to access objects in the R script. Will track dependencies for objects used within Rmd file.
}

