% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_program_graph_2.R
\name{create_program_graph}
\alias{create_program_graph}
\title{Make plot of project programs only
Summarize all programs. Sync status is assessed and indicated.}
\usage{
create_program_graph(project.id)
}
\arguments{
\item{project.id}{Project id of program}
}
\value{
List of data.frame of programs vertices, data.frame of edges, ggplot ,rgrapher=igraph
}
\description{
Make plot of project programs only
Summarize all programs. Sync status is assessed and indicated.
}
\details{
Uses ggplot2
}
\examples{
\dontrun{
create_program_graph("adaprHome")
} 
 
}
