% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_program_graph.R
\name{create_program_graph_v2}
\alias{create_program_graph_v2}
\title{Make plot of project programs only
Summarize all programs.}
\usage{
create_program_graph_v2(project.id)
}
\arguments{
\item{project.id}{Project id of program}
}
\value{
List of data.frame of programs vertices, data.frame of edges, ggplot ,rgrapher=igraph
}
\description{
Make plot of project programs only
Summarize all programs.
}
\details{
Deprecated. See graph.project()
}
\examples{
\dontrun{
 create_program_graph_v2("adaprHome")
} 
 
}
