% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_source_file_dir.R
\name{create_source_file_dir}
\alias{create_source_file_dir}
\title{Create source file directories}
\usage{
create_source_file_dir(project.id0 = project.id, source.file0 = source.file,
  source.description = "")
}
\arguments{
\item{project.id0}{is the project id name string}

\item{source.file0}{is the filename of the source}

\item{description}{is the string description of what the source file does}

\item{project.path}{is the home directory of the project}

\item{git.path}{path to git version control command}
}
\value{
source_info list describing the project
}
\description{
Create source file directories
}
\details{
Intializes git for the project, adds program git tracking, and initializes dependency tracking
}

