% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_program_graph_2.R
\name{createProgramGraph}
\alias{createProgramGraph}
\title{Make plot of project programs only
Summarize all programs. Sync status is assessed and indicated.}
\usage{
createProgramGraph(project.id, testSync = TRUE)
}
\arguments{
\item{project.id}{Project id of program}

\item{testSync}{Logical to test synchronization status}
}
\value{
List of data.frame of programs vertices, data.frame of edges, ggplot ,rgrapher=igraph
}
\description{
Make plot of project programs only
Summarize all programs. Sync status is assessed and indicated.
}
\details{
Uses ggplot2
}
\examples{
\dontrun{
createProgramGraph("adaprHome")
} 
 
}
