% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{dataDir}
\alias{dataDir}
\title{Returns project's data directory, allows relative directories. Used within an R script.}
\usage{
dataDir(project.id = getProject())
}
\arguments{
\item{project.id}{project specifies which data directory}
}
\value{
path to data directory
}
\description{
Returns project's data directory, allows relative directories. Used within an R script.
}
\examples{
\dontrun{
dataDir(getProject())
} 
}
