% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_setup.R
\name{defaultAdaprSetup}
\alias{defaultAdaprSetup}
\title{Set up adapr 1st time}
\usage{
defaultAdaprSetup()
}
\description{
Set up adapr 1st time
}
\details{
Use on ADAPR start up. Requires RStudio to work. Will make project directories in computer Document directory and create adaprHome project.
}
\examples{
\dontrun{
# Requires pandoc location or RStudio
default.adapr.setup()

} 

}
