% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_class.R
\docType{class}
\name{dependency-class}
\alias{dependency-class}
\alias{dependency}
\title{Dependency class}
\description{
Dependency class
}
\section{Methods}{

\describe{
\item{\code{update(df.update)}}{Updates the dependency object with a read in or write out}
}}

