% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Digest.R
\name{Digest}
\alias{Digest}
\title{Digest files from (digest package)}
\usage{
Digest(object, algo = c("md5", "sha1", "crc32", "sha256", "sha512"),
  serialize = TRUE, file = FALSE, length = Inf, skip = "auto",
  ascii = FALSE, raw = FALSE)
}
\value{
The filehash
}
\description{
Compute file hash without checking file.access.
}
\details{
Uses digest from package "digest". Authors Dirk Eddelbuettel edd@debian.org for the R interface; Antoine Lucas for the integration of crc32; Jarek Tuszynski for the file-based operations; Henrik Bengtsson and Simon Urbanek for improved serialization patches; Christophe Devine for the hash function implementations for sha-1, sha-256 and md5; Jean-loup Gailly and Mark Adler for crc32; Hannes Muehleisen for the integration of sha-512; Jim Hester for the integration of xxhash32, xxhash64 and murmur32.
}

