% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witchDoc.R
\name{docIDborn}
\alias{docIDborn}
\title{Create creation DocID based on true random number}
\usage{
docIDborn(nhash = 5, gitTF = FALSE, gitRepo = getwd())
}
\arguments{
\item{nhash}{Number of hashes to avoid hash collision}

\item{gitTF}{Logical whether to include git information}

\item{gitRepo}{path to git repository (defaults to working directory)}
}
\value{
data.frame
}
\description{
Create creation DocID based on true random number
}
\examples{

# Requires internet connection to access GitHub.
docIDborn()

}
