% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.orchard.R
\name{fileinfo.project}
\alias{fileinfo.project}
\title{List project file information disk space, modification timespan, days inactive}
\usage{
fileinfo.project(project.id = list.projects()$project.id)
}
\arguments{
\item{project.id}{character vector of projects}
}
\value{
dataframe with project information
}
\description{
List project file information disk space, modification timespan, days inactive
}
\examples{
\dontrun{
fileinfo.project()
} 

}
