% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.orchard.R
\name{fileInfoProjects}
\alias{fileInfoProjects}
\title{List project file information disk space, modification timespan, days inactive}
\usage{
fileInfoProjects(project.id = listProjects()$project.id)
}
\arguments{
\item{project.id}{character vector of projects}
}
\value{
dataframe with project information
}
\description{
List project file information disk space, modification timespan, days inactive
}
\examples{
\dontrun{
fileInfoProjects()
} 

}
