% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/finalize_dependency.R
\name{finalize_dependency}
\alias{finalize_dependency}
\title{Writes dependency data to file in "Dependency" directory}
\usage{
finalize_dependency(source_info, write = TRUE)
}
\arguments{
\item{source_info}{The source information list that contains dependency object}

\item{write}{Logical indicated to write the dependency object}
}
\value{
dependency.object
}
\description{
Writes dependency data to file in "Dependency" directory
}
\details{
Operates git tracking of program and dependency file.
}

