% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_orchard.R
\name{first.project}
\alias{first.project}
\title{Create first project}
\usage{
first.project(project.path = "", publish.path = "")
}
\arguments{
\item{project.path}{Path where first project will go}

\item{publish.path}{Path to share project results}
}
\value{
logical for succesful creation or not
}
\description{
Create first project
}
\examples{
\dontrun{
 opt <- get_adapr_options()
 first.project(opt$project.path,opt$publish.path)
} 

}
