% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.fs.time.R
\name{get.fs.time}
\alias{get.fs.time}
\title{Retrun time of file system}
\usage{
get.fs.time()
}
\value{
The file system write time
}
\description{
Writes to temporary file and extracts mod time with file.info
}

