% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_install_library_file.R
\name{get.library}
\alias{get.library}
\title{Get library for a project}
\usage{
get.library(project.id = get.project())
}
\arguments{
\item{project.id}{character vector of project}
}
\value{
dataframe of libraries
}
\description{
Get library for a project
}
\examples{
\dontrun{
get.library("adaprHome")
} 

}
