% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_project_path.R
\name{get.project.path}
\alias{get.project.path}
\title{Given Project name, Return project directory}
\usage{
get.project.path(project_name = NULL)
}
\arguments{
\item{project_name}{is string with project name}
}
\value{
string containing project directory
}
\description{
Given Project name, Return project directory
}
\details{
Reads "~/ProjectPaths/projectid_2_diretory.csv" into dataframe
}

