% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_path.R
\name{get.project.publish.path}
\alias{get.project.publish.path}
\title{Given Project name, Return project publish directory}
\usage{
get.project.publish.path(project_name = NULL)
}
\arguments{
\item{project_name}{is string with project name}
}
\value{
string containing project directory
}
\description{
Given Project name, Return project publish directory
}
\details{
Reads "~/ProjectPaths/projectid_2_diretory.csv" into dataframe
}
\examples{
\dontrun{
get.project.publish.path("adaprHome")
} 

}
