% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.adapr.options.R
\name{get.project}
\alias{get.project}
\title{Returns the  adapr project in R option "adaprProject"}
\usage{
get.project()
}
\value{
Value is specified project or default project
}
\description{
Returns the  adapr project in R option "adaprProject"
}
\details{
Default is adaprHome. Returns default if project does not exist.
}
\examples{
\dontrun{
 get.project()
} 

}
