% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_project_swap_directory.R
\name{get.project.swap.directory}
\alias{get.project.swap.directory}
\title{Retrieve project swap directory}
\usage{
get.project.swap.directory(project_name = NULL)
}
\arguments{
\item{project_name}{Project ID}
}
\value{
file path for swap directory
}
\description{
Retrieve project swap directory
}

