% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishers.R
\name{get.pubresults}
\alias{get.pubresults}
\title{Read result filepaths to publish}
\usage{
get.pubresults(project.id = get.project())
}
\arguments{
\item{project.id}{Project to publish}
}
\value{
dataframe of files to publish
}
\description{
Read result filepaths to publish
}
\details{
File is in support directory/files_to_publish.csv
}
\examples{
\dontrun{
get.pubresults("adaprHome")
} 

}
