% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.adapr.options.R
\name{get_adapr_options}
\alias{get_adapr_options}
\title{Returns the primary hub file with project location and id information}
\usage{
get_adapr_options(setoptions = FALSE)
}
\arguments{
\item{setoptions}{Logical specifying Execute Options}
}
\value{
adaproptions
}
\description{
Returns the primary hub file with project location and id information
}
\examples{
\dontrun{
opt <- get_adapr_options()
print(opt)
} 

}
