% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptSubgraph.R
\name{getDepSubgraph}
\alias{getDepSubgraph}
\title{Produces subgraph of dependencies of R script}
\usage{
getDepSubgraph(rscript, project.id = getProject(), plotTF = FALSE)
}
\arguments{
\item{rscript}{R script name}

\item{project.id}{Project graph}

\item{plotTF}{logical to plot or not}
}
\value{
list with subgraph in igraph format, data frame format, and layout for plottingss
}
\description{
Produces subgraph of dependencies of R script
}
\details{
Only needs one or the other argument.
}
\examples{
\dontrun{
 subGraph <- scriptSubgraph(project.id=getProject())
 plot(subGraph[[1]],vertex.label=basename(igraph::V(subGraph[[1]])$name),layout=subGraph[[2]])
} 

}
