% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.file.info.R
\name{getFileInfo}
\alias{getFileInfo}
\title{Retrieve the file info for the file by name OR from the data subdirectory data}
\usage{
getFileInfo(source_info, data = "", file0 = "", path.grep = "")
}
\arguments{
\item{source_info}{Project information in list}

\item{data}{Data file name}

\item{file0}{Filename not in data}

\item{path.grep}{Path to search for file in}
}
\value{
File information list outcome of search
}
\description{
Retrieve the file info for the file by name OR from the data subdirectory data
}
\examples{
\dontrun{
source_info <- create_source_file_dir("adaprHome","tree_controller.R")
getFileInfo(source_info,file0="read_data.R")
} 

}
