% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.fs.time.R
\name{getFileSysTime}
\alias{getFileSysTime}
\title{Retrun time of file system}
\usage{
getFileSysTime(directory = "")
}
\arguments{
\item{directory}{path to directorly}
}
\value{
The file system write time
}
\description{
Writes to temporary file and extracts mod time with file.info
}
\examples{
\dontrun{
getFileSysTime() 
} 

}
