% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_install_library_file.R
\name{getLibrary}
\alias{getLibrary}
\title{Get library for a project}
\usage{
getLibrary(project.id = getProject())
}
\arguments{
\item{project.id}{character vector of project}
}
\value{
dataframe of libraries
}
\description{
Get library for a project
}
\examples{
\dontrun{
getLibrary("adaprHome")
} 

}
