% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setProject.R
\name{getProject}
\alias{getProject}
\title{Returns the  adapr project in R option "adaprProject"}
\usage{
getProject()
}
\value{
Value is specified project or default project
}
\description{
Returns the  adapr project in R option "adaprProject"
}
\details{
Default is adaprHome. Returns default if project does not exist.
}
\examples{
\dontrun{
 getProject()
} 

}
