% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaceOperations.R
\name{getProjectFiles}
\alias{getProjectFiles}
\title{Returns all input and output files in a project.}
\usage{
getProjectFiles(project.id = getProject())
}
\arguments{
\item{project.id}{is the project id to lo}
}
\value{
data.frame with information about project's I/O files
}
\description{
Returns all input and output files in a project.
}
