% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_info_si.R
\name{getProjectInfoSI}
\alias{getProjectInfoSI}
\title{Given source_info object, retrieves project information}
\usage{
getProjectInfoSI(source_info)
}
\arguments{
\item{source_info}{is list with source information}
}
\value{
list with stacked dependency files, graph of dependencies, and condensed file information
}
\description{
Given source_info object, retrieves project information
}
\examples{
\dontrun{
source_info <- create_source_file_dir("adaprHome","tree_controller.R")
getProjectInfoSI(source_info)
} 

}
