% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_path.R
\name{getProjectPath}
\alias{getProjectPath}
\title{Given Project name, Return project directory}
\usage{
getProjectPath(project_name = NULL)
}
\arguments{
\item{project_name}{is string with project name}
}
\value{
string containing project directory
}
\description{
Given Project name, Return project directory
}
\details{
Reads "~/ProjectPaths/projectid_2_diretory.csv" into dataframe
}
\examples{
\dontrun{
getProjectPath("adaprHome")
} 

}
