% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_swap_directory.R
\name{getProjectSwapPath}
\alias{getProjectSwapPath}
\title{Retrieve project swap directory}
\usage{
getProjectSwapPath(project_name = NULL)
}
\arguments{
\item{project_name}{Project ID}
}
\value{
file path for swap directory
}
\description{
Retrieve project swap directory
}
\details{
Do not use. Deprecated.
}
