% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{getSourceInfo}
\alias{getSourceInfo}
\title{Returns the information related to the adapr script}
\usage{
getSourceInfo()
}
\value{
list with information about the project
}
\description{
Returns the information related to the adapr script
}
\examples{
\dontrun{
getSourceInfo()
} 

}
