% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptSubgraph.R
\name{getUpstream}
\alias{getUpstream}
\title{Identifies dependencies in a DAG}
\usage{
getUpstream(endVertex, isg)
}
\arguments{
\item{endVertex}{vertex}

\item{isg}{Project graph}
}
\value{
list with subgraph in igraph format, data frame format, and layout for plottingss
}
\description{
Identifies dependencies in a DAG
}
\details{
Lower level function. Uses recursion, may contain non-unique vertices.
}
