% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/git_add.R
\name{git.add}
\alias{git.add}
\title{git add to stage the file}
\usage{
git.add(gitdir, filename, branch = NULL, git_args = character(),
  git_binary = NULL)
}
\arguments{
\item{gitdir}{string with git directory}

\item{filename}{string of file to query}

\item{branch}{git branch}

\item{git_args}{string argument for git}

\item{git_binary}{location of git executable}
}
\value{
git output from git add
}
\description{
git add to stage the file
}

