% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/git.commit.R
\name{git.commit}
\alias{git.commit}
\title{git commit}
\usage{
git.commit(gitdir, message, branch = NULL, git_args = character(),
  git_binary = NULL)
}
\arguments{
\item{gitdir}{string with git directory}

\item{message}{for commit}

\item{branch}{git branch}

\item{git_args}{arguments to git}

\item{git_binary}{location of git executable}
}
\value{
git commit output
}
\description{
git commit
}

