% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_info.R
\name{git.info}
\alias{git.info}
\title{Retrieves the information from git about a file}
\usage{
git.info(gitdir, filename, branch = NULL, git_args = character(),
  git_binary = NULL)
}
\arguments{
\item{gitdir}{string with git directory}

\item{filename}{string of file to query}

\item{branch}{git branch}

\item{git_args}{string argument for git}

\item{git_binary}{location of git executable}
}
\value{
git log for filename
}
\description{
Retrieves the information from git about a file
}
\examples{
\dontrun{
si <- pull_source_info("adaprHome")
file0 <- file.path(si$project.path,project.directory.tree$analysis,"read_data.R")
git.info(si$project.path,file0) 
} 


}
