% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/git_init.R
\name{git.init}
\alias{git.init}
\title{Initiate git}
\usage{
git.init(gitdir, branch = NULL, git_binary = NULL)
}
\arguments{
\item{gitdir}{string with git directory}

\item{branch}{git branch}

\item{git_binary}{location of git executable}
}
\value{
git init lockaout
}
\description{
Initiate git
}

