% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_add.R
\name{gitAdd}
\alias{gitAdd}
\title{git add to stage the file}
\usage{
gitAdd(gitdir, filename, branch = NULL, git_args = character(),
  git_binary = NULL)
}
\arguments{
\item{gitdir}{string with git directory}

\item{filename}{string of file to query}

\item{branch}{git branch}

\item{git_args}{string argument for git}

\item{git_binary}{location of git executable}
}
\value{
git output from git add
}
\description{
git add to stage the file
}
\examples{
\dontrun{
si <- pullSourceInfo("adaprHome")
file0 <- file.path(si$project.path,project.directory.tree$analysis,"read_data.R")
gitAdd(si$project.path,file0) 
} 

}
\author{
Uses git2r package.
}
