% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.commit.R
\name{gitCommit}
\alias{gitCommit}
\title{git commit. Requires git installation.}
\usage{
gitCommit(gitdir, message, branch = NULL, git_args = character(),
  git_binary = NULL)
}
\arguments{
\item{gitdir}{string with git directory}

\item{message}{for commit}

\item{branch}{git branch}

\item{git_args}{arguments to git}

\item{git_binary}{location of git executable}
}
\value{
Not for direct use. Git commit output.
}
\description{
git commit. Requires git installation.
}
\examples{
\dontrun{
 si <- pullSourceInfo("adaprHome")
gitCommit(si$project.path,"test commit")
} 

}
