% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_configure.R
\name{gitConfigure}
\alias{gitConfigure}
\title{Configure user.name and email for git. Requires git installation.}
\usage{
gitConfigure(user.name, user.email)
}
\arguments{
\item{user.name}{Chris Someone}

\item{user.email}{someone[at]somewhere.com}
}
\value{
output from git
}
\description{
Configure user.name and email for git. Requires git installation.
}
\examples{
\dontrun{
# Requires Git installation
 gitConfigure("jonG","gelfond@somewhere.com")
} 

}
