% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_config_test.R
\name{gitConfigureTest}
\alias{gitConfigureTest}
\title{Checks git configuration. Requires git installation}
\usage{
gitConfigureTest()
}
\value{
integer 0 if configured otherwise 99 if not
}
\description{
Checks git configuration. Requires git installation
}
\examples{
\dontrun{
#Requires git installation
gitConfigureTest()
} 


}
\author{
Uses git_path adapted form devtools author Hadley Wickham
}
