% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitIngoreLibrary.R
\name{gitIgnoreLibrary}
\alias{gitIgnoreLibrary}
\title{Git ignore the library file}
\usage{
gitIgnoreLibrary(project.id = getProject())
}
\arguments{
\item{project.id}{to ignore Project's library}
}
\value{
success
}
\description{
Git ignore the library file
}
\details{
Libraries can be large and difficult to track with Git so we can ignore these.
}
