% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witchDoc.R
\name{gitSummary}
\alias{gitSummary}
\title{Create creation Git summary string with git2r function}
\usage{
gitSummary(gitRepo = "", gitTF = TRUE)
}
\arguments{
\item{gitRepo}{path to git repository (defaults to working directory)}

\item{gitTF}{logical to return git (if TRUE) or empty string (if FALSE)}
}
\value{
character string for git summary
}
\description{
Create creation Git summary string with git2r function
}
\examples{

# Requires internet connection to access random number generator.
\dontrun{
gitSummary()
}
}
