% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/guress_read_fcn.R
\name{guess.read.fcn}
\alias{guess.read.fcn}
\title{Return function for reading common file types}
\usage{
guess.read.fcn(filename)
}
\arguments{
\item{filename}{for file to be read}
}
\value{
function for reading
}
\description{
Return function for reading common file types
}
\details{
Uses the file suffix to return csv, read.delim, read.xls
}

