% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/guess_write_fnc.R
\name{guess.write.fcn}
\alias{guess.write.fcn}
\title{Return function for writing common file types}
\usage{
guess.write.fcn(filename)
}
\arguments{
\item{filename}{file to be written}
}
\value{
function for writing file
}
\description{
Return function for writing common file types
}
\details{
Uses the file suffix to return write.csv, png, pdf, save
}

