% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Harvest_trees.R
\name{Harvest.trees}
\alias{Harvest.trees}
\title{Collect trees from dependency directory}
\usage{
Harvest.trees(dependency.dir)
}
\arguments{
\item{dependency.dir}{Directory with dependency files}
}
\value{
data frame of stacked dependency files
}
\description{
Collect trees from dependency directory
}
\details{
adapts to relative or absolute directories
}
\examples{
\dontrun{
Harvest.trees(pull_source_info("adaprHome")$dependency.dir) 
} 


}
