% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashcopy.R
\name{hashcopy}
\alias{hashcopy}
\title{Given Project name, Return project publish directory}
\usage{
hashcopy(originals, targetdir)
}
\arguments{
\item{originals}{character vector of file paths to copy}

\item{targetdir}{directory to copy to}
}
\value{
vector of logicals indicating copy status
}
\description{
Given Project name, Return project publish directory
}
\details{
Only copies if there is a discrepancy in the file hash
}
