% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/id_new_libs.R
\name{id_new_libs}
\alias{id_new_libs}
\title{Captures libraries that are not loaded automatically}
\usage{
id_new_libs(library.data.file)
}
\arguments{
\item{library.data.file}{CSV File with a set of library names and repository locations}

\item{verbose}{Print which libraries are installed and loaded}
}
\value{
Libraries loaded that were not automatically loaded
}
\description{
Captures libraries that are not loaded automatically
}
\details{
Captures unaccounted for library within library information file
}

