% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_new_packages.R
\name{idPackages}
\alias{idPackages}
\title{Lower level function that captures packages that are explicitly loaded, not loaded automatically}
\usage{
idPackages(library.data.file)
}
\arguments{
\item{library.data.file}{CSV File with a set of library names and repository locations}
}
\value{
Libraries loaded that were not automatically loaded
}
\description{
Lower level function that captures packages that are explicitly loaded, not loaded automatically
}
\details{
Captures unaccounted for library within library information file. Not for direct use.
}
