% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importData}
\alias{importData}
\title{Import selected file into the project data directory}
\usage{
importData(datafile = "", project.id = getProject(), overwriteTF = TRUE)
}
\arguments{
\item{datafile}{filename can be unspecified}

\item{project.id}{string indicating which project data directory to copy into}

\item{overwriteTF}{logical indicating whether to overwrite exist file.}
}
\description{
Import selected file into the project data directory
}
\details{
If datafile is "" then a file choose dialogue is created
}
\examples{
\dontrun{
# Will open file browser to copy into adaprHome data directory
importData(project.id="adaprHome")

} 
}
