% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{importProject}
\alias{importProject}
\title{Imports new project using adapr import file or by selecting any R script within the project}
\usage{
importProject(importer.file = NULL, import.by.Rscript = FALSE)
}
\arguments{
\item{importer.file}{File created by adapr::makeImporter that is selected interactively OR any R script within program directory}

\item{import.by.Rscript}{logical indicating whether importer.file parameter is a "adapr_importer.R" file OR any R script within program directory.}
}
\value{
Logical indicating success or not
}
\description{
Imports new project using adapr import file or by selecting any R script within the project
}
\details{
This is a console command. The function adapr::makeImproter will create the "adapr_importer.R" file needed. Function will create an interactive file choosing window.
}
\examples{
\dontrun{
importProject() # Will open window for file selection.
} 
}
