% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_tree.R
\name{init.project}
\alias{init.project}
\title{initialize project}
\usage{
init.project(project.id, project.path = NA, publish.directory = NA)
}
\arguments{
\item{project.id}{Project name, if missing then default}

\item{project.path}{Project home directory, if missing then default}

\item{publish.directory}{Project branch exchange directory}
}
\value{
logical for success or not
}
\description{
initialize project
}
\details{
Wrapper for plant.tree
}
\examples{
\dontrun{
init.project("adaprTest")
} 
}

