% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intialize.dependency.info (Autosaved).R, R/intialize.dependency.info.R
\name{initialize_dependency_info}
\alias{initialize_dependency_info}
\title{Initializes dependency object source_info}
\usage{
initialize_dependency_info(source_info_arg)

initialize_dependency_info(source_info_arg)
}
\arguments{
\item{source_info_arg}{is a source_info list with describing R script and project}

\item{source_info_arg}{is a source_info list with describing R script and project}
}
\value{
Depedency file location

Depedency file location
}
\description{
Initializes dependency object source_info

Initializes dependency object source_info
}

