% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_install_library_file.R
\name{install}
\alias{install}
\title{Install package of specific version}
\usage{
install(package, version = NULL, installVersion = FALSE,
  lib = .libPaths()[1], repos = getOption("repos"),
  show.available = FALSE, packageSource = "", ...)
}
\arguments{
\item{package}{package name to install}

\item{version}{package version}

\item{installVersion}{logical, TRUE for install a specific version, if FALSE then latest}

\item{lib}{path to local library}

\item{repos}{character of repository}

\item{show.available}{logical to display whether the package is available}

\item{packageSource}{character describing where the papckage is from.}

\item{...}{Argument to install.packages/install.version functions}
}
\value{
Library information data
}
\description{
Install package of specific version
}
\details{
Installs from CRAN and bioconductor packages. Local libraries will not be installed.
}
