% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_library.R
\name{install_project_packages}
\alias{install_project_packages}
\title{Installs all packages}
\usage{
install_project_packages(project.id = get.project())
}
\arguments{
\item{project.id}{project.id to install packages for R package to}
}
\value{
Library information data
}
\description{
Installs all packages
}
\details{
Installs packages
}
\examples{
\dontrun{
install_project_packages("adaprHome")
} 
}
