% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_install_library_file.R
\name{installLibrary}
\alias{installLibrary}
\title{Install package of specific version}
\usage{
installLibrary(input = getLibrary(), lib = getProjectLibrary(),
  versionCheck = FALSE)
}
\arguments{
\item{input}{data.frame with 3 columns package version repos to install}

\item{lib}{path to local library}

\item{versionCheck}{logical to install specific version}
}
\value{
Library information data
}
\description{
Install package of specific version
}
\details{
Calls adapr::install and installs from CRAN and bioconductor packages. Local packages will not be installed.
}
\examples{
\dontrun{
setProject("adaprHome")
installLibrary()
} 

}
