% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_library.R
\name{installProjectPackages}
\alias{installProjectPackages}
\title{Installs all packages}
\usage{
installProjectPackages(project.id = getProject())
}
\arguments{
\item{project.id}{project.id to install packages for R package to}
}
\value{
Library information data
}
\description{
Installs all packages
}
\details{
Not for direct use. Installs autoloaded packages.
}
\examples{
\dontrun{
installProjectPackages("adaprHome")
} 
}
