% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Library.R
\name{Library}
\alias{Library}
\title{Installs and loads library specific to a project}
\usage{
Library(package, repository = "cran", github = "",
  project.id = getProject())
}
\arguments{
\item{package}{character for package to load/install}

\item{repository}{character for location of repository. "cran" for CRAN and "bioc" for bioconductor.}

\item{github}{character for devtools::install_github("xxx")}

\item{project.id}{project.id project id install within}
}
\description{
Installs and loads library specific to a project
}
\details{
Use within program Body
}
\examples{
\dontrun{
Library("adapr","cran",project.id="adaprHome")
} 

}
