% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list.branches}
\alias{list.branches}
\title{Lists the branches available for loading in the adapr project}
\usage{
list.branches(project.id = get.project())
}
\arguments{
\item{project.id}{project to find branches within}
}
\value{
dataframe of descriptions available branches
}
\description{
Lists the branches available for loading in the adapr project
}
\examples{
\dontrun{
list.branches("adaprHome")
} 


}
