% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list.programs}
\alias{list.programs}
\title{Lists the R scripts in the adapr project}
\usage{
list.programs(project.id = get.project())
}
\arguments{
\item{project.id}{project.id}
}
\value{
dataframe of R scripts and descriptions
}
\description{
Lists the R scripts in the adapr project
}
\examples{
\dontrun{
list.programs("adaprHome")
} 

}
