% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.orchard.R
\name{list.projects}
\alias{list.projects}
\title{List projects}
\usage{
list.projects(allInfo = TRUE)
}
\arguments{
\item{allInfo}{logical whether to return dat}
}
\value{
orchard
}
\description{
List projects
}
\examples{
\dontrun{
list.projects(TRUE)
} 


}
