% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{listDatafiles}
\alias{listDatafiles}
\title{Lists the branches available for loading in the adapr project}
\usage{
listDatafiles(si = source_info)
}
\arguments{
\item{si}{is source_info object}
}
\value{
description of data files
}
\description{
Lists the branches available for loading in the adapr project
}

