% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.orchard.R
\name{listProjects}
\alias{listProjects}
\title{List projects}
\usage{
listProjects(allInfo = TRUE)
}
\arguments{
\item{allInfo}{logical whether to return dat}
}
\value{
orchard
}
\description{
List projects
}
\examples{
\dontrun{
listProjects(TRUE)
} 


}
