% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{listScripts}
\alias{listScripts}
\title{Lists the R scripts in the adapr project}
\usage{
listScripts(project.id = getProject())
}
\arguments{
\item{project.id}{project.id}
}
\value{
dataframe of R scripts, descriptions, run time duration, last run/modification date, number of lines
}
\description{
Lists the R scripts in the adapr project
}
\examples{
\dontrun{
listScripts("adaprHome")
} 


}
