% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaceOperations.R
\name{listWorkspaces}
\alias{listWorkspaces}
\title{Lists the R script workspaces in the project's Results directory that are available for loading.}
\usage{
listWorkspaces(project = getProject())
}
\arguments{
\item{project}{character project ID to list the workspaces from}
}
\value{
data.frame with information about project's workspace files
}
\description{
Lists the R script workspaces in the project's Results directory that are available for loading.
}
\examples{
\dontrun{
setProject("adaprHome")
listWorkspaces()
} 
 
}
