% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_branch.R
\name{loadFlex}
\alias{loadFlex}
\title{Loads a single R object from file, more flexible than Load.branch or base::load}
\usage{
loadFlex(file, read.fcn = readRDS, ...)
}
\arguments{
\item{file}{contains R object}

\item{read.fcn}{function to read the file, default readRDS}

\item{...}{arguments passed to read.fcn}
}
\value{
object for file that was read
}
\description{
Loads a single R object from file, more flexible than Load.branch or base::load
}
\examples{
\dontrun{
processed <- load.flex("read_data.R/process_data.RData")
} 
}
