% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_install_library_file.R
\name{loadInstallLibraryFile}
\alias{loadInstallLibraryFile}
\title{This function is no longer supported. Loads libraries within the file library.list.file}
\usage{
loadInstallLibraryFile(library.data.file = NA, subgroup = NULL,
  verbose = FALSE, install.all = FALSE)
}
\arguments{
\item{library.data.file}{CSV File with a set of library names and repository locations}

\item{subgroup}{data frame with Package, repos, and specific columns}

\item{verbose}{Print which libraries are installed and loaded}

\item{install.all}{logicial indicated whether to install all project packages}
}
\value{
Library information data
}
\description{
This function is no longer supported. Loads libraries within the file library.list.file
}
\details{
Installs and loads all packages.Not for direct use. See installProjectPackages().
}
