% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaceOperations.R
\name{loadWorkspace}
\alias{loadWorkspace}
\title{Loads the R script workspace int}
\usage{
loadWorkspace(file = NA, script = getSourceInfo()$file$file,
  project = getProject())
}
\arguments{
\item{file}{character string of workspace file to load. If missing then loads last workspace of current R script.}

\item{script}{character string of R script to load the workspace from. By default loads the current R script's workspace.}

\item{project}{character project ID to load the workspaces from}
}
\value{
output from base::load of workspace file
}
\description{
Loads the R script workspace int
}
\details{
Used within an adapr R script or in the console. If file is stated then script parameter is ignored.
}
\examples{
\dontrun{
# To load a dumped workspace
loadWorkspace(script="mylastRscript.R",project="myProjectID")
# To load a preserved workspace from another R script
loadWorkspace(file="myFavoriteObjects.ws")
} 
 
}
