% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_dependency_graph.R
\name{Make.dependency.graph.obj}
\alias{Make.dependency.graph.obj}
\title{Creates an graph object from a dependency object}
\usage{
Make.dependency.graph.obj(dependency.out)
}
\arguments{
\item{dependency.out}{Dependency object(s) to make graph out of}
}
\value{
graph object of project/program dependencies
}
\description{
Creates an graph object from a dependency object
}
\examples{
\dontrun{
trees <- Harvest.trees(pull_source_info("adaprHome")$dependency.dir)
dag<-Make.dependency.graph.obj(trees)
} 
}
