% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprout_program.R
\name{make.program}
\alias{make.program}
\title{Generates the shell of a code that is project specific}
\usage{
make.program(project.id = get.project(), r = "", description = "",
  seed = 2011, run = TRUE)
}
\arguments{
\item{project.id}{Name of project}

\item{r}{is source file name or Filename to create}

\item{description}{What program does}

\item{seed}{Random start seed}

\item{run}{Execute r script?}
}
\value{
Logical indicating failure or not
}
\description{
Generates the shell of a code that is project specific
}
\details{
Will not overwrite existing program. Executes program stub. Mostly wrapper for sprout.program.
}
\examples{
\dontrun{
 make.program("adaprHome","read_data.R")
} 
}

