% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_summary_graph.R
\name{Make.summary.graph}
\alias{Make.summary.graph}
\title{Make.summary graph of projects based on files in dependency directory}
\usage{
Make.summary.graph(dependency.dir = NULL, dependency.object = NULL,
  plot.graph = FALSE)
}
\arguments{
\item{dependency.dir}{Dependency directory}

\item{dependency.object}{Dependency.data}

\item{plot.graph}{Logical to plot graph or not}
}
\description{
Make.summary graph of projects based on files in dependency directory
}
\details{
Only take dependency directory XOR dependency.data
}
\examples{
\dontrun{
trees <- Harvest.trees(pull_source_info("adaprHome")$dependency.dir)
dag<-Make.summary.graph(dependency.obj=trees)
plot(dag)
} 
}
