% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_dependency_graph.R
\name{makeDependencyGraphObj}
\alias{makeDependencyGraphObj}
\title{Creates an graph object from a dependency object}
\usage{
makeDependencyGraphObj(dependency.out)
}
\arguments{
\item{dependency.out}{Dependency object(s) to make graph out of}
}
\value{
graph object of project/program dependencies
}
\description{
Creates an graph object from a dependency object
}
\examples{
\dontrun{
trees <- readDependency(pullSourceInfo("adaprHome")$dependency.dir)
dag<-makeDependencyGraphObj(trees)
} 
}
