% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFunction.R
\name{makeFunction}
\alias{makeFunction}
\title{Generates the shell of a R function that is project specific in support_functions folder}
\usage{
makeFunction(functionName = NA, description = "",
  project.id = getProject())
}
\arguments{
\item{functionName}{character string for R function.}

\item{description}{character string description of function}

\item{project.id}{Name of project}
}
\value{
Logical indicating success or not
}
\description{
Generates the shell of a R function that is project specific in support_functions folder
}
\details{
Function file will add ".R" extension. Will not overwrite existing program.See makeScript() for making an R script.
}
