% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{makeImporter}
\alias{makeImporter}
\title{Creates an "adapr_importer.R" script that allows another user to import the project with 
the "importProject" function.}
\usage{
makeImporter(project.id = getProject())
}
\arguments{
\item{project.id}{name of project.}
}
\value{
output form base::write about success of file creation.
}
\description{
Creates an "adapr_importer.R" script that allows another user to import the project with 
the "importProject" function.
}
\details{
This is a console command. The function adapr::makeImproter will create the "adapr_importer.R" file needed by adapr::importProject fuction.
}
\examples{
\dontrun{
makeImporter() # Will create file called "adapr_importer.R" in top level project directory.
} 
}
