% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelsync_project.R
\name{monitorParallelSync.project}
\alias{monitorParallelSync.project}
\title{Track parallelSync.project while in progress}
\usage{
monitorParallelSync.project(project.id = get.project(), check.interval = 5)
}
\arguments{
\item{project.id}{Project to synchronize.}

\item{check.interval}{how many seconds to delay until last check}
}
\value{
ggplot of project graph
}
\description{
Track parallelSync.project while in progress
}
\details{
Must use separate R process from parallelSync.project(). Refreshes project plot with compute node labels are working or completed
}
\examples{
\dontrun{
monitorParallelSync.project("adaprHome")
} 

}
