% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelsync_project.R
\name{monitorParallelSync}
\alias{monitorParallelSync}
\title{Experimental (use with caution) track parallelSync while in progress}
\usage{
monitorParallelSync(project.id = getProject(), check.interval = 5)
}
\arguments{
\item{project.id}{Project to synchronize.}

\item{check.interval}{how many seconds to delay until last check}
}
\value{
ggplot of project graph
}
\description{
Experimental (use with caution) track parallelSync while in progress
}
\details{
Must use separate R process from parallelSync(). Refreshes project plot with compute node labels are working or completed
}
\examples{
\dontrun{
monitorParallelSync("adaprHome")
} 

}
