% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witchDoc.R
\name{newlinesDocID}
\alias{newlinesDocID}
\title{Generate string vector to render creation docID}
\usage{
newlinesDocID(bornTF = TRUE, ...)
}
\arguments{
\item{bornTF}{Logical to print the born on/create time (TRUE) or the compute time (FALSE-default)}

\item{...}{extra arguments to docIDborn or computed to git repository}
}
\value{
Character vector for docID
}
\description{
Generate string vector to render creation docID
}
\examples{

# Requires internet connection to access random number generator.
\dontrun{
newlinesDocID()
}

}
