% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.orchard.R
\name{open_orchard}
\alias{open_orchard}
\title{Browses orchard in file system}
\usage{
open_orchard()
}
\value{
orchard
}
\description{
Browses orchard in file system
}
\examples{
\dontrun{
open_orchard()
} 


}
