% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openScript.R
\name{openMarkdown}
\alias{openMarkdown}
\title{Opens R markdown from a project with default R program.}
\usage{
openMarkdown(markdownFile = getSourceInfo()$rmdfile$file,
  project.id = getProject())
}
\arguments{
\item{markdownFile}{Mardown filename to open, then current R script}

\item{project.id}{string for project id to search within.}
}
\description{
Opens R markdown from a project with default R program.
}
\details{
If markdownFile argument is blank or does not exist, then lists available scripts for convenience.
}
\examples{
\dontrun{
# Opens read_data.R within the adaprHome project
openMarkdown("read_data.Rmd","adaprHome")

} 

}
