% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.orchard.R
\name{openProjectList}
\alias{openProjectList}
\title{Browses orchard in file system}
\usage{
openProjectList()
}
\value{
orchard
}
\description{
Browses orchard in file system
}
\examples{
\dontrun{
openProjectList()
} 


}
