% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openScript.R
\name{openScript}
\alias{openScript}
\title{Opens script from a project with default R program. Can open markdown files as well.}
\usage{
openScript(rscript = "", project.id = getProject())
}
\arguments{
\item{rscript}{R script or Mardown filename to open}

\item{project.id}{string for project id to search within.}
}
\description{
Opens script from a project with default R program. Can open markdown files as well.
}
\details{
If rscript argument is blank, then lists available scripts for convenience.
}
\examples{
\dontrun{
# Opens read_data.R within the adaprHome project
openScript("read_data.R","adaprHome")

} 

}
