% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelsync_project.R
\name{parallelSync.project}
\alias{parallelSync.project}
\title{Uses DOSNOW package for parallel syncrhonization of project. Takes advantage of directed acyclic graph structure to run R script processes in parallel.}
\usage{
parallelSync.project(project.id = get.project(), n.cores = 2)
}
\arguments{
\item{project.id}{Project to synchronize.}

\item{n.cores}{Number of cores to use. Should be >1, but less than number of logical CPUs.}
}
\value{
data.frame with success/failure status.
}
\description{
Uses DOSNOW package for parallel syncrhonization of project. Takes advantage of directed acyclic graph structure to run R script processes in parallel.
}
\details{
Experimental. See also monitorParallelSync.project(), sync.project() and synctest.project(). Uses Results/tree_controller.R directory to pass work/completion data between nodes.
}
\examples{
\dontrun{
parallelsync.project("adaprHome")
} 


}
