% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelsync_project.R
\name{parallelSync}
\alias{parallelSync}
\title{Experimental (use with caution) parallel syncrhonization of project. Takes advantage of directed acyclic graph structure to run R script processes in parallel.}
\usage{
parallelSync(project.id = getProject(), n.cores = 2)
}
\arguments{
\item{project.id}{Project to synchronize.}

\item{n.cores}{Number of cores to use. Should be >1, but less than number of logical CPUs.}
}
\value{
data.frame with success/failure status.
}
\description{
Experimental (use with caution) parallel syncrhonization of project. Takes advantage of directed acyclic graph structure to run R script processes in parallel.
}
\details{
Experimental. See also monitorParallelSync(), syncProject() and syncTestProject(). Uses Results/tree_controller.R directory to pass work/completion data between nodes.
}
\examples{
\dontrun{
parallelsyncProject("adaprHome")
} 


}
