% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_recieve.R
\name{Pass}
\alias{Pass}
\title{Simple way to send objects from one script to another.}
\usage{
Pass(x, description = "results")
}
\arguments{
\item{x}{R object to send}

\item{description}{String describing R object stored in tree}
}
\value{
Value of Write.
}
\description{
Simple way to send objects from one script to another.
}
\details{
Adds '.rda' to object name and writes to Result file. Works with 'Receive'.
}
\examples{
\dontrun{
# Requires internet connection to access GitHub.
Pass(myObject,"great result")
} 
}
