% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_expand.2.R
\name{path.expand.2}
\alias{path.expand.2}
\title{Location of options file and project listing. Swap / for \\ in path expand.}
\usage{
path.expand.2(x)
}
\arguments{
\item{x}{file path, could be relative or ~}
}
\value{
Full filepath to x
}
\description{
Location of options file and project listing. Swap / for \\ in path expand.
}
\details{
See path.expand() in base R. Can be controlled in R profile by adding adaprHomeDir option. See adaprHomeDir().
}
\examples{
\dontrun{
 path.expand.2("~")
} 
}
