% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_orchard.R
\name{plant.orchard}
\alias{plant.orchard}
\title{Create project hub files in root directory}
\usage{
plant.orchard()
}
\value{
logical for succesful creation or not
}
\description{
Create project hub files in root directory
}
\examples{
\dontrun{
 plant.orchard()
} 

}
