% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_orchard.R
\name{plantOrchard}
\alias{plantOrchard}
\title{Create project hub files in root directory}
\usage{
plantOrchard()
}
\value{
logical for succesful creation or not
}
\description{
Create project hub files in root directory
}
\examples{
\dontrun{
 plantOrchard()
} 

}
