% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_tree.R
\name{plantTree}
\alias{plantTree}
\title{initialize project}
\usage{
plantTree(project.id, project.path = NA, swap.directory = NA,
  first.program = "read_data.R", project.libraryTF = FALSE,
  library.path = "")
}
\arguments{
\item{project.id}{Project name, if missing then default}

\item{project.path}{Project home directory, if missing then default}

\item{swap.directory}{Project branch exchange directory}

\item{first.program}{Name of first program in project (read_data.R default)}

\item{project.libraryTF}{Logical to use a local (not default) library}

\item{library.path}{= path to local (not default) library}
}
\value{
logical for success or not
}
\description{
initialize project
}
\details{
Not for direct use. See initProject().
}
