% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witchDoc.R
\name{printDocId}
\alias{printDocId}
\title{Create character strings for DocID with system user and compute time}
\usage{
printDocId(bornTF = FALSE, ...)
}
\arguments{
\item{bornTF}{Logical to print the born on/create time (TRUE) or the compute time (FALSE-default)}

\item{...}{extra arguments to docIDborn or computed to git repository}
}
\value{
data.frame
}
\description{
Create character strings for DocID with system user and compute time
}
\examples{

# Requires internet connection to access true random numbers
printDocId()


}
