% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/program_io_table.R
\name{program.io.table}
\alias{program.io.table}
\title{Create program io table}
\usage{
program.io.table(dependency.out)
}
\arguments{
\item{dependency.out}{Tree of dependencies}
}
\value{
Matrix summarizing inputs and outputs
}
\description{
Create program io table
}
\details{
groups inputs and outputs
}

