% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program_io_table.R
\name{programIOTable}
\alias{programIOTable}
\title{Create program io table}
\usage{
programIOTable(dependency.out)
}
\arguments{
\item{dependency.out}{Tree of dependencies}
}
\value{
Matrix summarizing inputs and outputs
}
\description{
Create program io table
}
\details{
groups inputs and outputs

Not for direct use.
}
\examples{
\dontrun{
trees <- readDependency(pullSourceInfo("adaprHome")$dependency.dir)
programIOTable(trees)
} 
}
