% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_directory_tree.R
\docType{data}
\name{project.directory.tree}
\alias{project.directory.tree}
\title{Project directory tree structure
contains the relative directory structure
included analysis, data, texidr, dependency.dir, support functions, and library bank 
locations}
\format{An object of class \code{list} of length 7.}
\usage{
project.directory.tree
}
\description{
Project directory tree structure
contains the relative directory structure
included analysis, data, texidr, dependency.dir, support functions, and library bank 
locations
}
\keyword{datasets}
