% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/project_report.R
\name{project.report}
\alias{project.report}
\title{Make plot of network within html documents.
Summarize all programs.}
\usage{
project.report(graph.width = 960, graph.height = 500)
}
\arguments{
\item{graph.width}{Sankey Plot dimensions}

\item{graph.height}{Sankey Plot dimensions}
}
\description{
Make plot of network within html documents.
Summarize all programs.
}
\details{
Assumes source_info in workspace
}

