% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/project_report_send.R
\name{project_report_send}
\alias{project_report_send}
\title{Make plot of network within html documents.
Summarize all programs.
Make a readme file at top project directory
copy to target.directory
Uses pander and pandoc unlike project_report}
\usage{
project_report_send(target.directory = source_info$project.path, source_info,
  send.data = FALSE, graph.width = 960, graph.height = 500)
}
\arguments{
\item{target.directory}{Location to send project report}

\item{source_info}{Source information list}

\item{send.data}{Logical to send data directory or not}

\item{graph.width}{Sankey Plot dimensions}

\item{graph.height}{Sankey Plot dimensions}
}
\description{
Make plot of network within html documents.
Summarize all programs.
Make a readme file at top project directory
copy to target.directory
Uses pander and pandoc unlike project_report
}
\details{
Dose not assume source_info in workspace
}

