% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_report_send_rmd.R
\name{project_report_send_rmd}
\alias{project_report_send_rmd}
\title{Make plot of network within html documents.
Summarize all programs.
Make a readme file at top project directory
copy to target.directory
Uses pander and pandoc unlike project_report}
\usage{
project_report_send_rmd(target.directory = get("source_info")$project.path,
  si, send.data = FALSE, graph.width = 960, graph.height = 500)
}
\arguments{
\item{target.directory}{Location to send project report}

\item{si}{Source information list}

\item{send.data}{Logical to send data directory or not}

\item{graph.width}{Sankey Plot dimensions}

\item{graph.height}{Sankey Plot dimensions}
}
\description{
Make plot of network within html documents.
Summarize all programs.
Make a readme file at top project directory
copy to target.directory
Uses pander and pandoc unlike project_report
}
\details{
Not for direct use. Does not assume source_info in workspace
}
\examples{
\dontrun{
 source_info <- create_source_file_dir("adaprHome","tree_controller.R")
project_report_send_rmd(si=source_info)
} 

}
