% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/project_report_si.R, R/project_reporter.R
\name{project_reporter}
\alias{project_reporter}
\title{Make plot of network within html documents.
Summarize all programs.}
\usage{
project_reporter(source_info, graph.width = 960, graph.height = 500)

project_reporter(source_info, graph.width = 960, graph.height = 500)
}
\arguments{
\item{source_info}{Source information list}

\item{graph.width}{Sankey Plot dimensions}

\item{graph.height}{Sankey Plot dimensions}

\item{source_info}{Source information list}

\item{graph.width}{Sankey Plot dimensions}

\item{graph.height}{Sankey Plot dimensions}
}
\description{
Make plot of network within html documents.
Summarize all programs.

Make plot of network within html documents.
Summarize all programs.
}
\details{
Assumes source_info in workspace

Assumes source_info in workspace
}

