% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/project_report_pander.R
\name{project_reporter_pander}
\alias{project_reporter_pander}
\title{Make plot of network within html documents.
Summarize all programs.
Uses pander and pandoc unlike project_report}
\usage{
project_reporter_pander(source_info, graph.width = 960, graph.height = 500)
}
\arguments{
\item{source_info}{Source information list}

\item{graph.width}{Sankey Plot dimensions}

\item{graph.height}{Sankey Plot dimensions}
}
\description{
Make plot of network within html documents.
Summarize all programs.
Uses pander and pandoc unlike project_report
}
\details{
Dose not assume source_info in workspace
}

