% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/project_report_pander_test.R
\name{project_reporter_pander_0}
\alias{project_reporter_pander_0}
\title{Make plot of network within html documents.
Summarize all programs.}
\usage{
project_reporter_pander_0(source_info, graph.width = 960,
  graph.height = 500)
}
\description{
Make plot of network within html documents.
Summarize all programs.
}

