% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapidPlot.R
\name{rapidPlot}
\alias{rapidPlot}
\title{Make project graph with sync status already computed.}
\usage{
rapidPlot(previousGraph, project.id = getProject(), message = "Running")
}
\arguments{
\item{previousGraph}{output value of createProjectGraph}

\item{project.id}{character string of project id examined}

\item{message}{string subtitle added to graph}
}
\description{
Make project graph with sync status already computed.
}
\details{
Used by syncProject() to disply synchronization progress
}
\examples{
\dontrun{
# Requires pandoc location or RStudio
graphData <- graphProject("adaprHome")
rapidPlot(graphData)

} 


}
