% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_library.R
\name{read_library}
\alias{read_library}
\title{Read library file}
\usage{
read_library(project.id = get.project())
}
\arguments{
\item{project.id}{project.id to read library file}
}
\value{
Library information data
}
\description{
Read library file
}
\details{
uses read.csv
}
\examples{
\dontrun{
read_library("adaprHome")
} 

}
