% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Harvest_trees.R
\name{readDependency}
\alias{readDependency}
\title{Collect trees from dependency directory}
\usage{
readDependency(dependency.dir)
}
\arguments{
\item{dependency.dir}{Directory with dependency files}
}
\value{
data frame of stacked dependency files
}
\description{
Collect trees from dependency directory
}
\details{
adapts to relative or absolute directories
}
\examples{
\dontrun{
readDependency(pullSourceInfo("adaprHome")$dependency.dir) 
} 


}
