% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{ReadTrack}
\alias{ReadTrack}
\title{Tracks files that read by functions not in adapr and captures the file information within dependency object}
\usage{
ReadTrack(file.name = "data.csv", description = "Data file")
}
\arguments{
\item{file.name}{name of file}

\item{description}{description of data file}
}
\value{
Filepath of file to read
}
\description{
Tracks files that read by functions not in adapr and captures the file information within dependency object
}
\details{
Allows tracking of files read by other functions than Read. Assumes file is in project "Data" directory
}

