% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_tree.R
\name{redirectTree}
\alias{redirectTree}
\title{Changes project directory/publish directory or identifies imported project}
\usage{
redirectTree(project.id0, project.path = NA, swap.directory = NA)
}
\arguments{
\item{project.id0}{Project name}

\item{project.path}{Project home directory}

\item{swap.directory}{Project publish directory}
}
\value{
logical for success or not
}
\description{
Changes project directory/publish directory or identifies imported project
}
\details{
Not for direct use. See relocate.project
}
