% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_tree.R
\name{relocate.project}
\alias{relocate.project}
\title{Changes project directory/publish directory or identifies imported project}
\usage{
relocate.project(project.id0, project.path = NA, swap.directory = NA)
}
\arguments{
\item{project.id0}{Project name}

\item{project.path}{Project home directory}

\item{swap.directory}{Project publish directory}
}
\value{
logical for success or not
}
\description{
Changes project directory/publish directory or identifies imported project
}
\details{
Is wrapper for redirect.tree. Does not move the project only indicates new location.
}
\examples{
\dontrun{
relocate.project("adaprTest","mydirectory1","mydirectory2publish")
} 
}
