% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.orchard.R
\name{remove.project}
\alias{remove.project}
\title{Removes project from orchard, but doesn't delete project from file system}
\usage{
remove.project(project.id0)
}
\arguments{
\item{project.id0}{which project to remove from orchard}
}
\value{
orchard
}
\description{
Removes project from orchard, but doesn't delete project from file system
}
\examples{
\dontrun{
remove.project("adaprHome")
reidentify.project("adaprHome")
} 


}
