% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_library.R
\name{remove_package}
\alias{remove_package}
\title{Remove R package to a project}
\usage{
remove_package(project.id = get.project(), library.name)
}
\arguments{
\item{project.id}{project.id to add R package to}

\item{library.name}{R package name to add}
}
\value{
Library information data
}
\description{
Remove R package to a project
}
\details{
Remove line from file
}
\examples{
\dontrun{
remove_package("adaprHome","ggplot2")
} 

}
