% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_library.R
\name{removePackage}
\alias{removePackage}
\title{Remove R package to a project}
\usage{
removePackage(project.id = getProject(), library.name)
}
\arguments{
\item{project.id}{project.id to add R package to}

\item{library.name}{R package name to add}
}
\value{
Library information data
}
\description{
Remove R package to a project
}
\details{
Not for direct use. Remove line from autoloading packages file.
}
\examples{
\dontrun{
removePackage("adaprHome","ggplot2")
} 

}
