% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.orchard.R
\name{removeProject}
\alias{removeProject}
\title{Removes project from orchard, but doesn't delete project from file system}
\usage{
removeProject(project.id0)
}
\arguments{
\item{project.id0}{which project to remove from orchard}
}
\value{
Project listing data frame.
}
\description{
Removes project from orchard, but doesn't delete project from file system
}
\examples{
\dontrun{
removeProject("adaprHome")
relcateProject("adaprHome")
} 


}
