% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_report_markdown.R
\name{report.project}
\alias{report.project}
\title{Make plot of network within html documents.
Summarize all programs.}
\usage{
report.project(project.id = get.project(), graph.width = 960,
  graph.height = 500)
}
\arguments{
\item{project.id}{Source information list}

\item{graph.width}{Sankey Plot dimensions}

\item{graph.height}{Sankey Plot dimensions}
}
\value{
File path to report html file
}
\description{
Make plot of network within html documents.
Summarize all programs.
}
\details{
Dose not assume source_info in workspace
}
\examples{
\dontrun{
report.project("adaprHome")
}  
}
