% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{resultsDir}
\alias{resultsDir}
\title{Returns project's results directory, allows relative directories. Only used within an R script, after create_source_file_dir.}
\usage{
resultsDir(sourceInfo = getSourceInfo())
}
\arguments{
\item{sourceInfo}{R source_info list created by create_source_file_dir}
}
\value{
path to data directory
}
\description{
Returns project's results directory, allows relative directories. Only used within an R script, after create_source_file_dir.
}
\examples{
\dontrun{
resultsDir(getSourceInfo())
} 
}
