% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rework_project_path.R
\name{rework.project.path}
\alias{rework.project.path}
\title{Collects all trees in dependency.dir and changes the project path}
\usage{
rework.project.path(dependency.dir = get("source_info")$depedency.dir,
  new.path = get.project.path(get("source_info")$project.id))
}
\arguments{
\item{dependency.dir}{location of dependency files to rework}

\item{new.path}{file path for the new project path}
}
\description{
Collects all trees in dependency.dir and changes the project path
}
\details{
Not for direct use. Used with swapping branches
}
