% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_Program.R
\name{runScriptQuiet}
\alias{runScriptQuiet}
\title{Run an R script in the background using callr::rcmd_safe}
\usage{
runScriptQuiet(pathToScript, ...)
}
\arguments{
\item{pathToScript}{filepath to R script}

\item{...}{extra arguments to callr::rcmd_safe}
}
\value{
value from callr::rcmd_safe status of script execution
}
\description{
Run an R script in the background using callr::rcmd_safe
}
\examples{
\dontrun{
runScriptQuiet("pathToMyRscript.R")
} 

}
