% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_source_directory.R
\name{runSourceDirectory}
\alias{runSourceDirectory}
\title{Runs all source files within the directory source.directory}
\usage{
runSourceDirectory(source.directory)
}
\arguments{
\item{source.directory}{is a directory with R source files to load}
}
\value{
source file list
}
\description{
Runs all source files within the directory source.directory
}
\details{
Looks for files with .R or .r suffixes.
}
\examples{
\dontrun{
 path <- getProjectPath("adaprHome")
 path <- file.path(path,"Programs",project.directory.tree$support)
 runSourceDirectory(path)
} 
}
