% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_loader.R
\name{scriptLoader}
\alias{scriptLoader}
\title{Initializes dependency object source_info}
\usage{
scriptLoader(projectID, script)
}
\arguments{
\item{projectID}{is the project id}

\item{script}{is the filename for the R script loaded}
}
\value{
value 1 if success
}
\description{
Initializes dependency object source_info
}
