% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptSubgraph.R
\name{scriptSubgraph}
\alias{scriptSubgraph}
\title{Produces script only subgraph}
\usage{
scriptSubgraph(project.id = getProject(), plotTF = FALSE)
}
\arguments{
\item{project.id}{Project id}

\item{plotTF}{logical to plot subgraph}
}
\value{
list with subgraph in igraph format, layout for plottingss
}
\description{
Produces script only subgraph
}
\details{
For use with getDepSubgraph
}
\examples{
\dontrun{
 subGraph <- scriptSubgraph(project.id=getProject())
 plot(subGraph[[1]],vertex.label=basename(V(subGraph[[1]])$name),layout=subGraph[[2]])
} 

}
