% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchScripts.R
\name{searchScripts}
\alias{searchScripts}
\title{Searches R scripts and R markdown files within a project.}
\usage{
searchScripts(matcher, project.id = getProject(), ...)
}
\arguments{
\item{matcher}{string or regular expression to identify within R or R markdown files}

\item{project.id}{string for project id to search within.}

\item{...}{arguments to grep}
}
\value{
Data frame with file names and counts of lines with matches.
}
\description{
Searches R scripts and R markdown files within a project.
}
\details{
Uses grep. Counts lines with matches, but repeats within a line are not counted.
}
\examples{
\dontrun{
# Opens read_data.R within the adaprHome project
searchScripts("read_data.R","adaprHome")

} 

}
