% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/send_branch.R
\name{send.branch}
\alias{send.branch}
\title{Copy dependent programs to swap directory}
\usage{
send.branch(branch_cut, all = FALSE)
}
\arguments{
\item{branch_cut}{filename of the base of the branch to send}

\item{all}{logical indicating whether to send all branches in project}
}
\description{
Copy dependent programs to swap directory
}

