% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_branch.R
\name{sendBranch}
\alias{sendBranch}
\title{Copy dependent programs to swap directory}
\usage{
sendBranch(branch_cut, all = FALSE)
}
\arguments{
\item{branch_cut}{filename of the base of the branch to send}

\item{all}{logical indicating whether to send all branches in project}
}
\description{
Copy dependent programs to swap directory
}
