% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_branch_si.R
\name{sendBranchSI}
\alias{sendBranchSI}
\title{Copy dependent programs to swap directory}
\usage{
sendBranchSI(source_info, branch_cut, all = FALSE)
}
\arguments{
\item{source_info}{Project information list}

\item{branch_cut}{filename of the base of the branch to send}

\item{all}{logical indicating whether to send all branches in project}
}
\description{
Copy dependent programs to swap directory
}
