% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.adapr.options.R
\name{set.project}
\alias{set.project}
\title{Checks or changes the specified adapr project in R option "adaprProject"}
\usage{
set.project(project.id = "", quickTest = FALSE)
}
\arguments{
\item{project.id}{characters specifying project.id of working project}

\item{quickTest}{logical whether to check if project exists}
}
\value{
value is specified project or default project
}
\description{
Checks or changes the specified adapr project in R option "adaprProject"
}
\details{
Default is adaprHome. Returns default if project does not exist.
}
\examples{
\dontrun{
 set.project("adaprHome")
} 

}
