% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.adapr.options.R
\name{set_adapr_options}
\alias{set_adapr_options}
\title{Returns Modifies the primary adapr option file}
\usage{
set_adapr_options(optionname = "", optionvalue = "")
}
\arguments{
\item{optionname}{is name of option to modify}

\item{optionvalue}{is new value to give optionname}
}
\value{
adaproptions
}
\description{
Returns Modifies the primary adapr option file
}
\examples{
\dontrun{
opt <- get_adapr_options()
set_adapr_options("project.path",opt$project.path)
opt2 <- get_adapr_options()
identical(opt,opt2)
} 

}
