% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.adapr.options.R
\name{setAdaprOptions}
\alias{setAdaprOptions}
\title{Returns Modifies the primary adapr option file}
\usage{
setAdaprOptions(optionname = "", optionvalue = "")
}
\arguments{
\item{optionname}{is name of option to modify}

\item{optionvalue}{is new value to give optionname}
}
\value{
adaproptions
}
\description{
Returns Modifies the primary adapr option file
}
\examples{
\dontrun{
opt <- getAdaprOptions()
setAdaprOptions("project.path",opt$project.path)
opt2 <- getAdaprOptions()
identical(opt,opt2)
} 

}
