% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setProject.R
\name{setProject}
\alias{setProject}
\title{Checks or changes the specified adapr project in R option "adaprProject"}
\usage{
setProject(project.id = "", quickTest = TRUE)
}
\arguments{
\item{project.id}{characters specifying project.id of working project}

\item{quickTest}{logical whether to check if project exists}
}
\value{
value is specified project or default project
}
\description{
Checks or changes the specified adapr project in R option "adaprProject"
}
\details{
Default is adaprHome. Returns default if project does not exist.
}
\examples{
\dontrun{
 setProject("adaprHome")
} 

}
