% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{show.project}
\alias{show.project}
\title{Opens project directory}
\usage{
show.project(project.id = get.project())
}
\arguments{
\item{project.id}{character string specifies project to open}
}
\description{
Opens project directory
}
\details{
Use BrowseURL to open project directory.
}
\examples{
\dontrun{
show.project("adaprHome")
} 

}
