% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{show.results}
\alias{show.results}
\title{Opens results directory of project or R script within a project}
\usage{
show.results(project.id = get.project(), rscript = "")
}
\arguments{
\item{project.id}{character string specifies project}

\item{rscript}{character string specifies the R script result directory to open}
}
\description{
Opens results directory of project or R script within a project
}
\details{
Use BrowseURL to open results directory
}
\examples{
\dontrun{
show.results("adaprHome")
} 

}
