% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{showProject}
\alias{showProject}
\title{Opens project directory}
\usage{
showProject(si = source_info)
}
\arguments{
\item{si}{is source_info object}
}
\description{
Opens project directory
}
\details{
Use BrowseURL to open project directory
}

