% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{showResults}
\alias{showResults}
\title{Opens results directory}
\usage{
showResults(si = source_info)
}
\arguments{
\item{si}{is source_info object}
}
\description{
Opens results directory
}
\details{
Use BrowseURL to open results directory
}

