% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_sync_si_load.R
\name{source_sync_si_load}
\alias{source_sync_si_load}
\title{Synchronize project by IDENTIFYING necessary R scripts}
\usage{
source_sync_si_load(source_info)
}
\arguments{
\item{source_info}{Project information within source_info list}
}
\value{
Data.frame with sources needed to synchronize with run times
}
\description{
Synchronize project by IDENTIFYING necessary R scripts
}
\details{
Not usually direct use. See sync.project() and synctest.project().
}
\examples{
\dontrun{
source_info <- create_source_file_dir("adaprHome","tree_controller.R")
source_sync_si_load(source_info)
} 
}
