% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witchDoc.R
\name{sourceFinder}
\alias{sourceFinder}
\title{sourceFinder creates a filename and hash at runtime}
\usage{
sourceFinder(parentFile0 = parent.frame(2)$ofile)
}
\arguments{
\item{parentFile0}{path to parent file that creates docutment}
}
\value{
data.frame with parent file name and its hash (sha1)
}
\description{
sourceFinder creates a filename and hash at runtime
}
\examples{

# Requires internet connection to access random number generator.
\dontrun{
sourceFinder(parentFile0=parent.frame(2)$ofile)
}
}
