% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_sync_si.R
\name{sourceSyncSI}
\alias{sourceSyncSI}
\title{Lower level function that synchronizes project by running necessary R scripts. Loads from source_info list.}
\usage{
sourceSyncSI(source_info, run = TRUE, plot.to.file = FALSE)
}
\arguments{
\item{source_info}{Project information within source_info list}

\item{run}{logical indicated whether to run or just identify asynchrony}

\item{plot.to.file}{logical for writing file in tree_controller.R directory}
}
\value{
Data.frame with sources needed to synchronize with run times
}
\description{
Lower level function that synchronizes project by running necessary R scripts. Loads from source_info list.
}
\details{
Not usually direct use. See syncProject() and syncTestProject().
}
\examples{
\dontrun{
source_info <- create_source_file_dir("adaprHome","tree_controller.R")
sourceSyncSI(source_info)
} 
}
