% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sprout_program.R
\name{sprout.program}
\alias{sprout.program}
\title{Generates the shell of a code that is project specific}
\usage{
sprout.program(project.id = NA, source.file.name = NA, description = "",
  seed = 2011, capture.load.command = "library(IT2)", controller = FALSE)
}
\arguments{
\item{project.id}{Name of project}

\item{source.file.name}{Filename to create}

\item{description}{What program does}

\item{seed}{Set seed at program initialization}

\item{capture.load.command}{Command for loading inference tree library}
}
\value{
Logical indicating success or not
}
\description{
Generates the shell of a code that is project specific
}
\details{
Will not overwrite existing program
}

