% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_project_paths.R
\name{swapProjectPath}
\alias{swapProjectPath}
\title{Lower level function that takes list of dependency file data and changes the project path}
\usage{
swapProjectPath(list.deps,
  new.path = getProjectPath(get("source_info")$project.id))
}
\arguments{
\item{list.deps}{list of dependency file data}

\item{new.path}{file path for the new project path}
}
\value{
Updated list of dependency data
}
\description{
Lower level function that takes list of dependency file data and changes the project path
}
\details{
Not for direct use. Used with swapping branches by reworkProjectPath()
}
