% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_test_si.R
\name{sync.project}
\alias{sync.project}
\title{Checks the synchronization project and runs scripts needed for synchronization}
\usage{
sync.project(project.id)
}
\arguments{
\item{project.id}{is project to synchronize}
}
\value{
Character string with message about success for synchronization
}
\description{
Checks the synchronization project and runs scripts needed for synchronization
}
\examples{
\dontrun{
sync.project("adaprHome")
} 



}

