% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_test.R
\name{Sync.test.OLD}
\alias{Sync.test.OLD}
\title{check the synchrony of source files and their created objects}
\usage{
Sync.test.OLD(dagger, tree, plotl = TRUE)
}
\arguments{
\item{dagger}{a directed acyclic igraph representing dependencies}

\item{tree}{dependency tree corresponding to dagger}

\item{plotl}{logical for plotting or not}
}
\description{
check the synchrony of source files and their created objects
}
