% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sync_test_pi.R
\name{Sync.test.pi}
\alias{Sync.test.pi}
\title{Tests the synchrony of files in dependency tree}
\usage{
Sync.test.pi(project_info, plotl0 = FALSE)
}
\arguments{
\item{project_info}{Project information from get_project_info function}

\item{plotl0}{Logical indicated whether to plot the updated files}
}
\value{
list or logical indicated whether project is synchronized or not
}
\description{
Tests the synchrony of files in dependency tree
}
\details{
Not for direct use. See sync.project() and synctest.project().
}
\examples{
\dontrun{
si <- pull_source_info("adaprHome")
projInfo  <- get.project.info(si$dependency.dir)
Sync.test.pi(projInfo)
} 


}
