% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_test_si.R
\name{synctest.project}
\alias{synctest.project}
\title{Tests the synchrony of files in dependency tree}
\usage{
synctest.project(project.id = get.project())
}
\arguments{
\item{project.id}{is project to test the synchrony of}
}
\value{
list with logical indicated whether project is synchronized or not and details about synchrony
}
\description{
Tests the synchrony of files in dependency tree
}
\examples{
\dontrun{
synctest.project("adaprHome")
} 


}
