% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_test_New.R
\name{syncTest}
\alias{syncTest}
\title{Lower level function that checks the synchrony of source files and their created objects}
\usage{
syncTest(dagger, tree, plotl = FALSE)
}
\arguments{
\item{dagger}{a directed acyclic igraph representing dependencies}

\item{tree}{dependency tree corresponding to dagger}

\item{plotl}{logical for plotting or not}
}
\value{
list with synchronizing information
}
\description{
Lower level function that checks the synchrony of source files and their created objects
}
\details{
Not for direct use. See syncProject() and syncTestProject().
}
\examples{
\dontrun{
 si <- pullSourceInfo("adaprHome")
projInfo <- getProjectInfo(si$dependency.dir)
syncTest(projInfo$graph,projInfo$tree)
} 
}
