% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sync_test_pi.R
\name{syncTestPI}
\alias{syncTestPI}
\title{Lower level function that tests the synchrony of files in dependency tree given project information.}
\usage{
syncTestPI(project_info, plotl0 = FALSE)
}
\arguments{
\item{project_info}{Project information from get_project_info function}

\item{plotl0}{Logical indicated whether to plot the updated files}
}
\value{
list or logical indicated whether project is synchronized or not
}
\description{
Lower level function that tests the synchrony of files in dependency tree given project information.
}
\details{
Not for direct use. See syncProject() and syncTestProject().
}
\examples{
\dontrun{
si <- pullSourceInfo("adaprHome")
projInfo  <- getProjectInfo(si$dependency.dir)
syncTestPI(projInfo)
} 


}
