% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_test_si.R
\name{syncTestProject}
\alias{syncTestProject}
\title{Tests the synchrony of files in dependency tree}
\usage{
syncTestProject(project.id = getProject())
}
\arguments{
\item{project.id}{is project to test the synchrony of}
}
\value{
list with logical indicated whether project is synchronized or not and details about synchrony
}
\description{
Tests the synchrony of files in dependency tree
}
\examples{
\dontrun{
syncTestProject("adaprHome")
} 


}
