% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_test_si.R
\name{syncTrunk}
\alias{syncTrunk}
\title{Partial project synchronization of dependencies to an Rscript. Runs only scripts needed for synchronization.}
\usage{
syncTrunk(rscript, project.id = getProject(), ask = FALSE)
}
\arguments{
\item{rscript}{script to synchronize the output of}

\item{project.id}{is project to synchronize}

\item{ask}{logical whether to report estimated run time prior to execution}
}
\value{
Character string with message about success for synchronization
}
\description{
Partial project synchronization of dependencies to an Rscript. Runs only scripts needed for synchronization.
}
\examples{
\dontrun{
syncTrunk("read_data.R","adaprHome")
} 



}
