% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAdaprConfig.R
\name{updateAdaprConfig}
\alias{updateAdaprConfig}
\title{Updates the project list file to include project specific libraries.}
\usage{
updateAdaprConfig()
}
\value{
orchard
}
\description{
Updates the project list file to include project specific libraries.
}
\details{
Adds 2 columns to project listing. (project.library and library.path). This enables project specific libraries.

Projects are by default set to use the default library. If packrat is used, this is probably not the case.
}
\examples{
\dontrun{
orchard <- updateOrchardLibraries()
print(subset(orchard,project.id=="adaprHome"))
} 

}
