% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.orchard.R
\name{updateOrchardLibraries}
\alias{updateOrchardLibraries}
\title{Updates the project list file to include project specific libraries.}
\usage{
updateOrchardLibraries()
}
\value{
orchard
}
\description{
Updates the project list file to include project specific libraries.
}
\details{
Adds 2 columns to project listing. (project.library and library.path). This enables project specific libraries.
}
\examples{
\dontrun{
orchard <- updateOrchardLibraries()
print(subset(orchard,project.id=="adaprHome"))
} 

}
