% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.orchard.R
\name{uproot_tree}
\alias{uproot_tree}
\title{Removes project from orchard, but doesn't delete project from file system}
\usage{
uproot_tree(project.id0)
}
\arguments{
\item{project.id0}{which project to remove from orchard}
}
\value{
orchard
}
\description{
Removes project from orchard, but doesn't delete project from file system
}
\examples{
\dontrun{
uproot_tree("adaprHome")
} 


}

