% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{viewBranches}
\alias{viewBranches}
\title{Opens available Branches as a csv file}
\usage{
viewBranches(projectId = getProject())
}
\arguments{
\item{projectId}{character string specifying project}
}
\value{
'listBranches' output
}
\description{
Opens available Branches as a csv file
}
\details{
Viewing only. Does not edit the data! Writes into the temp directory and uses system's .csv file browser.
}
\examples{
\dontrun{
viewData("adaprHome")
} 

}
