% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{viewData}
\alias{viewData}
\title{Opens a data frame as a csv file}
\usage{
viewData(df, overwriteTF = FALSE)
}
\arguments{
\item{df}{data frame to write and then open}

\item{overwriteTF}{logical to overwrite existing file. Default is FALSE}
}
\value{
path to data directory
}
\description{
Opens a data frame as a csv file
}
\details{
Viewing only. Does not edit the data! Writes into the temp directory and uses system's .csv file browser.
}
\examples{
\dontrun{
viewData(mtcars)
} 

}
