% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witchDoc.R
\name{witchUser}
\alias{witchUser}
\title{Retrieve or identify a username.}
\usage{
witchUser(userName = "Anonymous", setUser = FALSE)
}
\arguments{
\item{userName}{String to specify user's name}

\item{setUser}{logical to set user as userName}
}
\value{
character string for git summary
}
\description{
Retrieve or identify a username.
}
\details{
Accesses the options()$WitchUser variable or the system user name.
}
\examples{
\dontrun{
witchUser()
}
}
