% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dependency.R
\name{write.dependency}
\alias{write.dependency}
\title{Lower level function that writes the dependency object to file}
\usage{
write.dependency(dependency.object, dependency.pointer)
}
\arguments{
\item{dependency.object}{dependency object to output}

\item{dependency.pointer}{filename to write}
}
\value{
TRUE
}
\description{
Lower level function that writes the dependency object to file
}
\details{
Not for direct use. See finalize_dependency().
}
