% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write.R
\name{WriteTrack}
\alias{WriteTrack}
\title{Tracks files written by functions not in adapr and captures the file information within dependency object}
\usage{
WriteTrack(file.name = "data.csv", description = "Result file")
}
\arguments{
\item{file.name}{name of file}

\item{description}{description of data file}
}
\value{
Filepath of file that was written
}
\description{
Tracks files written by functions not in adapr and captures the file information within dependency object
}
\details{
Allows tracking of files written by other functions than Write. Assumes file is in Results directory
}
\examples{
\dontrun{
source_info <- create_source_file_dir("adaprHome","tree_controller.R")
write.csv(cars,file.path(source_info$results.dir,"test.csv"))
WriteTrack("cars.csv","cars dataframe")
showResults()  
} 
 
}
