% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaprSheet.R
\name{adaprSheet}
\alias{adaprSheet}
\title{Browse adapr cheat sheet}
\usage{
adaprSheet()
}
\description{
Browse adapr cheat sheet
}
\details{
Opens pdf from R package within system
}
\examples{
\dontrun{
# Requires pandoc location or RStudio
adaprSheet()

} 

}
