% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishers.R
\name{browsePubFiles}
\alias{browsePubFiles}
\title{Browses publication table for editing}
\usage{
browsePubFiles(project.id = getProject())
}
\arguments{
\item{project.id}{Project to publish}
}
\value{
dataframe of files to publish
}
\description{
Browses publication table for editing
}
\details{
File is in support directory/files_to_publish.csv
}
\examples{
\dontrun{
browsePubFiles("adaprHome")
} 


}
