% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_file_info.R
\name{Create.file.info}
\alias{Create.file.info}
\title{Creates a list containing information about a file}
\usage{
Create.file.info(path, file, description = "", dbname = "", si = NULL)
}
\arguments{
\item{path}{is the directory file resides}

\item{file}{is the filename}

\item{description}{is the description of the file}

\item{dbname}{is the database name}

\item{si}{is the source_info list default is null}
}
\value{
file info list with path filename fullname description and db.name
}
\description{
Creates a list containing information about a file
}

