% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_source_file_dir.R
\name{create_source_file_dir}
\alias{create_source_file_dir}
\title{Create source file directories}
\usage{
create_source_file_dir(project.id = "", source.file,
  source.description = "", project.path = get.project.path(project.id),
  git.path = NULL, git.log = TRUE)
}
\arguments{
\item{project.id}{is the project id name string}

\item{source.file}{is the filename of the source}

\item{project.path}{is the home directory of the project}

\item{git.path}{path to git version control command}

\item{description}{is the string description of what the source file does}
}
\value{
source_info list describing the project
}
\description{
Create source file directories
}
\details{
Intializes git for the project, adds program git tracking, and initializes dependency tracking
}

