% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_source_file_dir.R
\name{create_source_file_dir}
\alias{create_source_file_dir}
\title{Create source file directories}
\usage{
create_source_file_dir(project.id0 = get("project.id"),
  source.file0 = get("source.file"), source.description = "")
}
\arguments{
\item{project.id0}{project id name string}

\item{source.file0}{filename of the source}

\item{source.description}{character description of what the source file does}
}
\value{
source_info list describing the project
}
\description{
Create source file directories
}
\details{
Intializes git for the project, adds program git tracking, creates project library and initializes dependency tracking.
Creates directories for Project and Results. Initialize the file tracking object source_info.
Gathers file information on all project files.Initialize Git for project (if using Git).
Adds dependencies files to Git
Initialized archivist  repo for script 
Run all R scripts in support_function directory 
Run all R scripts in script specific support_function/myRscript.R directory 
Create R markdown file with same file prefix (if not already done).
Create publication file (if not already done).
}
\examples{
\dontrun{
source_info <- create_source_file_dir("adaprHome","tree_controller.R")
} 

}
