% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Digest.R
\name{Digest}
\alias{Digest}
\title{Digest files from (digest package)}
\usage{
Digest(object, algo = c("md5", "sha1", "crc32", "sha256", "sha512"),
  serialize = TRUE, file = FALSE, length = Inf, skip = "auto",
  ascii = FALSE, raw = FALSE)
}
\value{
The filehash
}
\description{
Compute file hash without checking file.access
}

