% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witchDoc.R
\name{docIDcomputed}
\alias{docIDcomputed}
\title{Create compute DocID with system user and compute time}
\usage{
docIDcomputed(gitTF = TRUE, gitRepo = getwd(),
  parentFile = parent.frame(2)$ofile)
}
\arguments{
\item{gitTF}{Logical whether to include git information}

\item{gitRepo}{path to git repository (defaults to working directory)}

\item{parentFile}{filename of source file}

\item{...}{arguments to sourceFinder}
}
\value{
data.frame
}
\description{
Create compute DocID with system user and compute time
}
\examples{

# Requires internet connection to access GitHub.
\dontrun{
docIDcomputed()
}
}
