% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_dependency.R
\name{finalize_dependency}
\alias{finalize_dependency}
\title{Writes dependency data to file in "Dependency" directory}
\usage{
finalize_dependency(RMD = FALSE, write = TRUE)
}
\arguments{
\item{write}{Logical indicated to write the dependency object}
}
\value{
dependency.object
}
\description{
Writes dependency data to file in "Dependency" directory
}
\details{
Operates git tracking of program and dependency file.

Strips project directory out of dependency file
}

