% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_dependency.R
\name{finalize_dependency}
\alias{finalize_dependency}
\title{Writes dependency data to file in "Dependency" directory}
\usage{
finalize_dependency(RMD = TRUE, write = TRUE)
}
\arguments{
\item{RMD}{Logical denoting whether the finalizing occurs in markdown vs R script}

\item{write}{Logical indicated to write the dependency object}
}
\value{
dependency.object
}
\description{
Writes dependency data to file in "Dependency" directory
}
\details{
Operates git tracking of program and dependency file.

Strips project directory out of dependency file.
Passes workspace to R markdown file and render.
Stores file modification and hash for all dependencies.
Writes dependency information to \file{Dependency} directory.
Adds dependency file and Session information to Git.
}
\examples{
\dontrun{
#Executed only at the end of adapr R script
#finalize_dependency() 
} 

}
