% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_project_info_si.R
\name{get.project.info.si}
\alias{get.project.info.si}
\title{Given source_info object, retrieves project information}
\usage{
get.project.info.si(source_info)
}
\arguments{
\item{source_info}{is list with source information}
}
\value{
list with stacked dependency files, graph of dependencies, and condensed file information
}
\description{
Given source_info object, retrieves project information
}

