% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.orchard.R
\name{get_orchard}
\alias{get_orchard}
\title{Returns the primary hub file with project location and id information}
\usage{
get_orchard()
}
\value{
orchard
}
\description{
Returns the primary hub file with project location and id information
}

