% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/git_configure.R
\name{git.configure}
\alias{git.configure}
\title{Configure user.name and email for git}
\usage{
git.configure(user.name, user.email)
}
\arguments{
\item{user.name}{Chris Someone}

\item{user.email}{someone[at]somewhere.com}
}
\value{
output from git
}
\description{
Configure user.name and email for git
}

