% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/git_config_test.R
\name{git.configure.test}
\alias{git.configure.test}
\title{Checks git configuration}
\usage{
git.configure.test()
}
\value{
integer 0 if configured otherwise 99 if not
}
\description{
Checks git configuration
}

