% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/git_path.R
\name{git_path}
\alias{git_path}
\title{Find path for git executable}
\usage{
git_path(git_binary_name = NULL)
}
\arguments{
\item{git_binary_name}{git binary name}
}
\value{
git executable
}
\description{
Find path for git executable
}

