% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_config_test.R
\name{gitConfigureTest}
\alias{gitConfigureTest}
\title{Checks git configuration. Requires git installation}
\usage{
gitConfigureTest(globalTF = TRUE)
}
\arguments{
\item{globalTF}{logical specifying global git configuration}
}
\value{
output from git2r::config
}
\description{
Checks git configuration. Requires git installation
}
\examples{
\dontrun{
#Requires git installation
gitConfigureTest()
} 


}
\author{
Uses git2r::config in git2r package.
}
