% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Graph.R
\name{Graph}
\alias{Graph}
\title{Write object and capture file information}
\usage{
Graph(file.name = "data.csv", description = "Result file",
  write.fcn = guess.write.fcn(file.name), date = FALSE, ...)
}
\arguments{
\item{file.name}{file to write to the source "Result" directory}

\item{description}{describes object to write}

\item{write.fcn}{function for writing file of object type}

\item{date}{logical for adding date to filename}
}
\value{
File information list
}
\description{
Write object and capture file information
}

