% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_tree.R
\name{init.project}
\alias{init.project}
\title{initialize project}
\usage{
init.project(project.id, project.path = NA, publish.directory = NA,
  first.program = "read_data.R")
}
\arguments{
\item{project.id}{Project name, if missing then default}

\item{project.path}{Project home directory, if missing then default}

\item{publish.directory}{Project branch exchange directory}

\item{first.program}{Name of first program in project (read_data.R default)}
}
\value{
logical for success or not
}
\description{
initialize project
}
\details{
Sets project

Wrapper for plant.tree
}
\examples{
\dontrun{
init.project("adaprTest")
} 
}
