% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intialize.dependency.info.R
\name{initialize_dependency_info}
\alias{initialize_dependency_info}
\title{Initializes dependency object source_info}
\usage{
initialize_dependency_info(source_info_arg)
}
\arguments{
\item{source_info_arg}{is a source_info list with describing R script and project}
}
\value{
Depedency file location
}
\description{
Initializes dependency object source_info
}
\details{
Not for direct use.
}
