% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{listBranches}
\alias{listBranches}
\title{Lists the branches available for loading in the adapr project}
\usage{
listBranches(project.id = getProject())
}
\arguments{
\item{project.id}{project to find branches within}
}
\value{
dataframe of descriptions available branches
}
\description{
Lists the branches available for loading in the adapr project
}
\examples{
\dontrun{
listBranches("adaprHome")
} 


}
