% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{listBranches}
\alias{listBranches}
\title{Lists the branches available for loading in the adapr project}
\usage{
listBranches(si = get("source_info"))
}
\arguments{
\item{si}{is source_info object}
}
\value{
dataframe of descriptions available branches
}
\description{
Lists the branches available for loading in the adapr project
}
\examples{
\dontrun{
source_info <- create_source_file_dir("adaprHome","tree_controller.R")
} 
}
