% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{listDatafiles}
\alias{listDatafiles}
\title{Lists the data files available for reading in the adapr project}
\usage{
listDatafiles(project.id = getProject())
}
\arguments{
\item{project.id}{Project to look for data files within}
}
\value{
description of data files
}
\description{
Lists the data files available for reading in the adapr project
}
\examples{
\dontrun{
listDatafiles("adaprHome")
} 

}
