% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{listDatafiles}
\alias{listDatafiles}
\title{Lists the data files available for reading in the adapr project}
\usage{
listDatafiles(si = get("source_info"))
}
\arguments{
\item{si}{is source_info object}
}
\value{
description of data files
}
\description{
Lists the data files available for reading in the adapr project
}
\details{
Deprecated. See list.datafiles()
}
