% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.orchard.R
\name{listProjects}
\alias{listProjects}
\title{List projects}
\usage{
listProjects(project.id0 = "", allInfo = TRUE)
}
\arguments{
\item{project.id0}{character for specific project. Empty string default will list all projects.}

\item{allInfo}{logical whether to return all data}
}
\value{
data frame with project information
}
\description{
List projects
}
\examples{
\dontrun{
listProjects(TRUE)
} 


}
