% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{listScripts}
\alias{listScripts}
\title{Lists the R scripts in the adapr project}
\usage{
listScripts(project.id = get("source_info")$project.id)
}
\arguments{
\item{project.id}{project.id}
}
\value{
dataframe of R scripts and descriptions
}
\description{
Lists the R scripts in the adapr project
}
\details{
Deprecated see list.programs
}
