% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_branch.R
\name{Load.branch}
\alias{Load.branch}
\title{Loads a single R object from file}
\usage{
Load.branch(file)
}
\arguments{
\item{file}{contains R object}
}
\value{
object for file that was read
}
\description{
Loads a single R object from file
}

