% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_source_directory.R
\name{load.source.directory}
\alias{load.source.directory}
\title{Runs all source files within the directory source.directory}
\usage{
load.source.directory(source.directory)
}
\arguments{
\item{source.directory}{is a directory with R source files to load}
}
\value{
source file list
}
\description{
Runs all source files within the directory source.directory
}
\details{
Looks for files with .R or .r suffixes.
}

