% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_setup.R
\name{loadAdaprTest}
\alias{loadAdaprTest}
\title{Create adaprTest example project}
\usage{
loadAdaprTest(localLibraryTF = FALSE, overwrite = TRUE)
}
\arguments{
\item{localLibraryTF}{Logical for local library or not}

\item{overwrite}{Logical indicating whether to overwrite existing project}
}
\description{
Create adaprTest example project
}
\details{
To be run after default adapr set up.
}
\examples{
\dontrun{
# 
 loadAdaprTest()

} 

}
