% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_hyperlink.R
\name{make.hyperlink}
\alias{make.hyperlink}
\title{Makes HTML hyper link}
\usage{
make.hyperlink(files, links)
}
\arguments{
\item{files}{character vector of filenames}
}
\value{
link command vector
}
\description{
Makes HTML hyper link
}
\details{
Used in making HTML files
}

