% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprout_program.R
\name{makeScript}
\alias{makeScript}
\title{Generates the shell of a code that is project specific}
\usage{
makeScript(r = "", description = "", project.id = getProject(),
  seed = 2011, run = TRUE, openTF = TRUE)
}
\arguments{
\item{r}{is source file name or Filename to create}

\item{description}{Character string describing what program does}

\item{project.id}{Character string for name of project}

\item{seed}{Random start seed}

\item{run}{Logical for execution of r script}

\item{openTF}{Logcial for opening R script}
}
\value{
Logical indicating failure or not
}
\description{
Generates the shell of a code that is project specific
}
\details{
Will not overwrite existing program. Executes program and opens stub program. Mostly wrapper for sproutProgram.
}
\examples{
\dontrun{
 makeScript("read_data.R",description="reads data","adaprHome")
} 
}
