% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plan_orchard.R
\name{plant.orchard}
\alias{plant.orchard}
\title{Create project hub files in root directory}
\usage{
plant.orchard()
}
\value{
logical for succesful creation or not
}
\description{
Create project hub files in root directory
}

