% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_tree.R
\name{plant.tree}
\alias{plant.tree}
\title{initialize project}
\usage{
plant.tree(project.id, project.path = NA, swap.directory = NA,
  first.program = "read_data.R")
}
\arguments{
\item{project.id}{Project name, if missing then default}

\item{project.path}{Project home directory, if missing then default}

\item{swap.directory}{Project branch exchange directory}

\item{first.program}{Name of first program in project (read_data.R default)}
}
\value{
logical for success or not
}
\description{
initialize project
}
\details{
Not for direct use. See init.project().
}
