% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plant_tree.R
\name{plant.tree}
\alias{plant.tree}
\title{initialize project}
\usage{
plant.tree(project.id, project.path, swap.directory)
}
\arguments{
\item{project.id}{Project name}

\item{project.path}{Project home directory}

\item{swap.directory}{Project branch exchange directory}
}
\value{
logical for success or not
}
\description{
initialize project
}

