% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/project_directory_tree.R
\docType{data}
\name{project.directory.tree}
\alias{project.directory.tree}
\title{Project directory tree structure
contains the relative directory structure
included analysis, data, texidr, dependency.dir, support functions, and library bank
locations}
\format{\preformatted{List of 7
 $ analysis      : chr "Programs"
 $ data          : chr "Data"
 $ results       : chr "Results"
 $ tex.dir       : chr "texdir"
 $ dependency.dir: chr "Programs/Dependency"
 $ support       : chr "Programs/support_functions"
 $ library.bank  : chr "Libraries"
}}
\usage{
project.directory.tree
}
\description{
Project directory tree structure
contains the relative directory structure
included analysis, data, texidr, dependency.dir, support functions, and library bank
locations
}
\keyword{datasets}

