% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_report_markdown.R
\name{project_report_markdown}
\alias{project_report_markdown}
\title{Make plot of network within html documents.
Summarize all programs.
Uses pandoc unlike project_report}
\usage{
project_report_markdown(source_info, graph.width = 960, graph.height = 500)
}
\arguments{
\item{source_info}{Source information list}

\item{graph.width}{Sankey Plot dimensions}

\item{graph.height}{Sankey Plot dimensions}
}
\description{
Make plot of network within html documents.
Summarize all programs.
Uses pandoc unlike project_report
}
\details{
Dose not assume source_info in workspace
}

