% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_source_info.R
\name{pull_source_info}
\alias{pull_source_info}
\title{Create source_info from project.id}
\usage{
pull_source_info(project.id)
}
\arguments{
\item{project.id}{Project ID to use}
}
\value{
source_info for generic source_info tree operations
}
\description{
Create source_info from project.id
}
\details{
Creates tree_controller.R directory. This directory is used for operation on the tree.
}
