% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_source_info.R
\name{pull_source_info}
\alias{pull_source_info}
\title{Create source_info from project.id}
\usage{
pull_source_info(project.id)
}
\arguments{
\item{project.id}{}
}
\value{
source_info for generic source_info tree operations
}
\description{
Create source_info from project.id
}
\details{
Creates tree_controller.R directory
}

