% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Read_cap.R
\name{Read.cap}
\alias{Read.cap}
\title{Read data and capture the file information within dependency object}
\usage{
Read.cap(file.info, read.fcn, source_info, ...)
}
\arguments{
\item{file.info}{file information list}

\item{read.fcn}{function for reading the file}

\item{source_info}{source information list}
}
\value{
object read from files
}
\description{
Read data and capture the file information within dependency object
}

