% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_dependency.R
\name{read.dependency}
\alias{read.dependency}
\title{Read the dependcy data from file}
\usage{
read.dependency(dependency.pointer)
}
\arguments{
\item{dependency.pointer}{filename with dependency data}
}
\value{
dependency data.frame
}
\description{
Read the dependcy data from file
}

