% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_library.R
\name{readLibrary}
\alias{readLibrary}
\title{Read library file}
\usage{
readLibrary(project.id = getProject())
}
\arguments{
\item{project.id}{project.id to read library file}
}
\value{
Autoloaded library information data
}
\description{
Read library file
}
\details{
Not for direct use. Uses read.csv to read autoloaded libraries.
}
\examples{
\dontrun{
readLibrary("adaprHome")
} 

}
