% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_tree.R
\name{relocateProject}
\alias{relocateProject}
\title{changes project directory/publish directory/library locataion or identifies imported project}
\usage{
relocateProject(project.id0, project.path = NA, swap.directory = NA,
  project.libraryTF = FALSE, library.path = "")
}
\arguments{
\item{project.id0}{Project name}

\item{project.path}{Project home directory}

\item{swap.directory}{Project publish directory}

\item{project.libraryTF}{Logical to use a local (not default) library}

\item{library.path}{= path to local (not default) library}
}
\value{
logical for success or not
}
\description{
changes project directory/publish directory/library locataion or identifies imported project
}
\details{
Is wrapper for redirectTree. Does not move the project only indicates new location.
}
\examples{
\dontrun{
relocateProject("adaprTest","mydirectory1","mydirectory2publish")
} 
}
