% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Render.R
\name{Render_Rmd}
\alias{Render_Rmd}
\title{Renders and Rmarkdown file}
\usage{
Render_Rmd(Rmd.file, description = "Rmarkdown", ...)
}
\arguments{
\item{Rmd.file}{Filename of .Rmd file within the R scripts' results directory (source_info$results.dir)}

\item{...}{extra arguments for rmarkdown::render}
}
\value{
Rendered output file information
}
\description{
Renders and Rmarkdown file
}
\details{
Uses rmarkdown library to access objects in the R script. Will track dependencies for objects used within Rmd file.
}

