% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rework_project_path.R
\name{rework.project.path}
\alias{rework.project.path}
\title{Collects all trees in dependency.dir and changes the project path}
\usage{
rework.project.path(dependency.dir = source_info$depedency.dir,
  new.path = get.project.path(source_info$project.id))
}
\arguments{
\item{dependency.dir}{location of dependency files to rework}

\item{new.path}{file path for the new project path}
}
\description{
Collects all trees in dependency.dir and changes the project path
}
\details{
Used with swapping branches
}

