% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_Program.R
\name{run.program}
\alias{run.program}
\title{Run an R script within a project using devtools::clean_source}
\usage{
run.program(r = get("source_info")$file$file, project.id = get.project(),
  logRmd = FALSE)
}
\arguments{
\item{r}{R script within that project (r is short R script for convenience)}

\item{project.id}{project id}

\item{logRmd}{logical indicating whether to create R markdown log}
}
\value{
value from clean_source from devtools package
}
\description{
Run an R script within a project using devtools::clean_source
}
\examples{
\dontrun{
run.program("read_data.R","adaprHome")
} 

}
