% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{showProject}
\alias{showProject}
\title{Opens project directory}
\usage{
showProject(si = get("source_info"), project.id = "")
}
\arguments{
\item{si}{is source_info object}

\item{project.id}{character string specifies project to open}
}
\description{
Opens project directory
}
\details{
Use BrowseURL to open project directory
}
