% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{showResults}
\alias{showResults}
\title{Opens results directory}
\usage{
showResults(si = get("source_info"), project.id = "", rscript = "")
}
\arguments{
\item{si}{is source_info object}

\item{project.id}{character string specifies project}

\item{rscript}{character string specifies the R script result directory to open}
}
\description{
Opens results directory
}
\details{
Deprecated: use show.results(). Uses BrowseURL to open results directory
}
