% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_test_si.R
\name{sync.project}
\alias{sync.project}
\title{Checks the synchronization project and runs scripts needed for synchronization}
\usage{
sync.project(project.id = get.project(), ask = FALSE)
}
\arguments{
\item{project.id}{is project to synchronize}

\item{ask}{logical whether to report estimated run time prior to execution}
}
\value{
Character string with message about success for synchronization
}
\description{
Checks the synchronization project and runs scripts needed for synchronization
}
\examples{
\dontrun{
sync.project("adaprHome")
} 



}
