% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_test_3.R
\name{Sync.test}
\alias{Sync.test}
\title{check the synchrony of source files and their created objects}
\usage{
Sync.test(dagger, tree, plotl = FALSE)
}
\arguments{
\item{dagger}{a directed acyclic igraph representing dependencies}

\item{tree}{dependency tree corresponding to dagger}

\item{plotl}{logical for plotting or not}
}
\value{
list with synchronizing information
}
\description{
check the synchrony of source files and their created objects
}
\details{
Not for direct use. See sync.project() and synctest.project().
}
\examples{
\dontrun{
 si <- pull_source_info("adaprHome")
projInfo <- get.project.info(si$dependency.dir)
Sync.test(projInfo$graph,projInfo$tree)
} 
}
