% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sync_test.R
\name{Sync.test}
\alias{Sync.test}
\title{check the synchrony of source files and their created objects}
\usage{
Sync.test(dagger, tree, plotl = TRUE)
}
\arguments{
\item{tree}{dependency tree corresponding to dagger}

\item{plotl}{logical for plotting or not}

\item{g.all}{dependency object or set of dependency objects making a directed acyclic graph}
}
\description{
check the synchrony of source files and their created objects
}

