% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sync_test_si.R
\name{sync.test.si}
\alias{sync.test.si}
\title{Tests the synchrony of files in dependency tree}
\usage{
sync.test.si(source_info, plotl0 = FALSE)
}
\arguments{
\item{source_info}{source_info containing project information}

\item{plotl0}{Logical indicated whether to plot the updated files}
}
\value{
list or logical indicated whether project is synchronized or not
}
\description{
Tests the synchrony of files in dependency tree
}

