% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_test_2.R
\name{Sync.test_2nd}
\alias{Sync.test_2nd}
\title{check the synchrony of source files and their created objects}
\usage{
Sync.test_2nd(dagger, tree, plotl = TRUE)
}
\arguments{
\item{dagger}{a directed acyclic igraph representing dependencies}

\item{tree}{dependency tree corresponding to dagger}

\item{plotl}{logical for plotting or not}
}
\value{
list with synchronizing information
}
\description{
check the synchrony of source files and their created objects
}
