% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_cap.R
\name{Write.cap}
\alias{Write.cap}
\title{Write data and capture the file information within dependency object}
\usage{
Write.cap(obj = NULL, file.info, write.fcn, source_info, ...)
}
\arguments{
\item{obj}{object to write, if null then open graphics device}

\item{file.info}{file information list}

\item{source_info}{source information list}
}
\value{
file.info file information returned
}
\description{
Write data and capture the file information within dependency object
}

