% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_dependency.R
\name{write.dependency}
\alias{write.dependency}
\title{Write the dependcy object to file}
\usage{
write.dependency(dependency.object, dependency.pointer)
}
\arguments{
\item{dependency.object}{dependency object to output}

\item{dependency.pointer}{filename to write}
}
\value{
TRUE
}
\description{
Write the dependcy object to file
}

